% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{all.bundles.to.data.frames}
\alias{all.bundles.to.data.frames}
\title{all.bundles.to.data.frames}
\usage{
\method{all}{bundles.to.data.frames}(bundles, design, sep = "›")
}
\arguments{
\item{bundles}{a list of xml text files representing the fhir bundles.}

\item{design}{a structure that specifies which table should contain which entries of the bundle.}

\item{sep}{a string to separate pasted multiple entries}
}
\value{
a list of data frames.
}
\description{
converts all fhir bundles to a list of data frames
}
\examples{
\dontrun{
all.bundles.to.data.frames( bundles, design )
}
}
