% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{bundle.tag.values}
\alias{bundle.tag.values}
\title{bundle.tag.values}
\usage{
bundle.tag.values(bundle, xpath)
}
\arguments{
\item{bundle}{fhir bundle}

\item{xpath}{the path to the value in the xml document.}
}
\value{
a single value or vector
}
\description{
extracts values from bundle root path.
}
\examples{
\dontrun{
bundle.tag.values( bundle, xpath = "total" )
}
}
