% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bundle2df}
\alias{bundle2df}
\title{Extracts one data frame out of one bundle}
\usage{
bundle2df(bundle, design.df, sep = " -+- ", add_ids = F)
}
\arguments{
\item{bundle}{A xml object containing one FHIR bundle}

\item{design.df}{On element of the design from \code{\link{fhir_crack}}, i.e. a list of length 1
or 2, where the first element is a XPath expression to the ressource and the (optional)
second element is either a XPath expression or a named list containing column names and XPath expressions}

\item{sep}{A string to separate pasted multiple entries.}

\item{add_ids}{Logical Scalar. Should indices be added to multiple entries?}
}
\description{
Extracts one data frame out of one bundle
}
\examples{
#unserialize example bundle
bundles <- lapply(medication_bundles, xml2::xml_unserialize)

#extract first bundle
bundle <- bundles[[1]]

#define design
design <- list(
     ".//MedicationStatement",
     list(
	      SYSTEM  = "medicationCodeableConcept/coding/system/@value",
	      CODE    = "medicationCodeableConcept/coding/code/@value",
	      DISPLAY = "medicationCodeableConcept/coding/display/@value"
	      )
	 )


#convert bundle to data frame
result <- fhircrackr:::bundle2df(bundle, design)
}
