% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{bundle2df_auto}
\alias{bundle2df_auto}
\title{Flatten a xml node set automatically}
\usage{
bundle2df_auto(xml, xpath, sep = " -+- ", ns.id = NULL)
}
\arguments{
\item{xml}{A xml document.}

\item{xpath}{A xpath to the root node.}

\item{sep}{A separator for multiple entries.}

\item{ns.id}{A string containing the namespace id.}
}
\value{
A data frame containing the all nodes entries.
}
\description{
Flatten a xml node set automatically
}
\examples{
\dontrun{
bundle2df_auto(xml, xpath="//resource")
}
}
