% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{bundles.to.dataframes}
\alias{bundles.to.dataframes}
\title{bundles.to.dataframes}
\usage{
bundles.to.dataframes(bundles, design)
}
\arguments{
\item{bundles}{a list of xml text files representing the fhir bundles.}

\item{design}{a structure that specifies which table should contain which entries of the bundle.}
}
\value{
a list of data frames.
}
\description{
converts all fhir bundles to a list of data frames
}
\examples{
\dontrun{
bundles.to.dataframes( bundles, design )
}
}
