% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bundles2df}
\alias{bundles2df}
\title{Convert several bundles to one data frame}
\usage{
bundles2df(
  bundles,
  design.df,
  sep = " -+- ",
  add_indices = F,
  brackets = c("<", ">")
)
}
\arguments{
\item{bundles}{A list of xml objects containing FHIR bundles}

\item{design.df}{On element of the design from \code{\link{fhir_crack}}, i.e. a list of length 1
or 2, where the first element is a XPath expression to the ressource and the (optional)
second element is either a XPath expression or a named list containing column names and XPath expressions}

\item{sep}{A string to separate pasted multiple entries.}

\item{add_indices}{A Logical Scalar.}

\item{brackets}{A Vector of Strings defining the Brackets surrounding the Indices. e.g. c( "<", ">")}
}
\description{
Convert several bundles to one data frame
}
\examples{
#unserialize example bundle
bundles <- fhir_unserialize(medication_bundles)

#define design
design <- list(
     ".//MedicationStatement",
     list(
	      SYSTEM  = "medicationCodeableConcept/coding/system/@value",
	      CODE    = "medicationCodeableConcept/coding/code/@value",
	      DISPLAY = "medicationCodeableConcept/coding/display/@value"
	      )
	 )


#convert bundles to data frame
result <- fhircrackr:::bundles2df(bundles, design)
}
