% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{bundles2dfs}
\alias{bundles2dfs}
\title{bundles2dfs}
\usage{
bundles2dfs(bundles, design, sep = "-+-")
}
\arguments{
\item{bundles}{a fhir search result as a list of xml text files.}

\item{design}{a structure that specifies which table should contain which entries of the bundle.}

\item{sep}{a string to separate pasted multiple entries}
}
\value{
a list of data frames.
}
\description{
converts all fhir bundles (the result of a get.bundles) to a list of data frames
}
\examples{
\dontrun{
bundles2dfs( bundles, design )
}
}
