% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{capability.statement}
\alias{capability.statement}
\title{capability.statement}
\usage{
capability.statement(
  url = "https://hapi.fhir.org/baseR4",
  sep = " -+- ",
  remove.empty.columns = T
)
}
\arguments{
\item{url}{the url of the fhir server endpoint.}

\item{sep}{a string to separate pasted multiple entries}
}
\value{
a data frame.
}
\description{
get the capability statement about a fhir server.
}
\examples{
\dontrun{
capability.statement( "https://hapi.fhir.org/baseR4" )
}
}
