% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{capability_statement}
\alias{capability_statement}
\title{Get capability statement}
\usage{
capability_statement(
  url = "https://hapi.fhir.org/baseR4",
  sep = " -+- ",
  remove.empty.columns = T
)
}
\arguments{
\item{url}{The url of the fhir server endpoint.}

\item{sep}{A string to separate pasted multiple entries}

\item{remove.empty.columns}{Logical Scalar. Remove empty columns?}
}
\value{
A data frame with the capability statement.
}
\description{
Get the capability statement about a fhir server.
}
\examples{
\dontrun{
capability_statement("https://hapi.fhir.org/baseR4")
}
}
