% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{coerce.types}
\alias{coerce.types}
\title{coerce.types}
\usage{
coerce.types(df, stringsAsFactors = F)
}
\arguments{
\item{df}{a data frame with strings as column entries}

\item{stringsAsFactors}{strings as factors}
}
\value{
a data frame with coerced types
}
\description{
coerce a data frame's columns
}
\examples{
\dontrun{
coerce.types( dfs$Besuch )
}
}
