% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{coerce_types}
\alias{coerce_types}
\title{Coerce columns}
\usage{
coerce_types(df, stringsAsFactors = F)
}
\arguments{
\item{df}{A data frame with strings as column entries.}

\item{stringsAsFactors}{Strings as factors.}
}
\value{
A data frame with coerced types.
}
\description{
Coerce a data frame's columns.
}
\examples{
\dontrun{
coerce_types(dfs$Besuch)
}
}
