% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{concatenate.paths}
\alias{concatenate.paths}
\title{concatenate.paths}
\usage{
concatenate.paths(path1 = "w", path2 = "d", os = "LiNuX")
}
\arguments{
\item{path1}{the left hand part of final path.}

\item{path2}{the right hand part of final path.}

\item{os}{the operating system you'r operating on: windows or linux.}
}
\value{
a string
}
\description{
concatenates to paths strings correctly
}
\examples{
\dontrun{
concatenate.paths( "data", "patients" )
concatenate.paths( "/data", "patients" )
concatenate.paths( "/data/", "patients" )
concatenate.paths( "/data", "/patients" )
concatenate.paths( "/data/", "/patients/" )
concatenate.paths( "data", "patients", "windows" )
}
}
