% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{conformance}
\alias{conformance}
\title{conformance}
\usage{
conformance(url)
}
\arguments{
\item{url}{the url of the fhir server}
}
\value{
a data frame.
}
\description{
get the conformance information of a fhir server.
}
\examples{
\dontrun{
conformance( "https://vonk.fire.ly/R4" )
}
}
