% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{crack}
\alias{crack}
\title{Crack it!}
\usage{
crack(bundles, design, sep = " -+- ", ns.id = NULL)
}
\arguments{
\item{bundles}{A list of fhir bundles as xml docs.}

\item{design}{A named list specifiying which data.frame should contain which entries of the bundle.
The names correspond to the names of the resulting data.frames. Each element of design is a list of length 2 where
the first element is a XPath expression to locate the entry in a fhir bundle page and second element is a named list
with XPath expressions of locations to the values of the items in the bundle page. The names of this second element named
list are the column names of the resulting data.frames.}

\item{sep}{A string to separate pasted multiple entries.}

\item{ns.id}{A string containing the namespace id.}
}
\value{
A list of data frames as specified by \code{design}
}
\description{
Crack it!
}
