% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{transaction_bundle_example}
\alias{transaction_bundle_example}
\alias{example_resource}
\title{Toy examples to POST on a server}
\format{
An object of class \code{fhir_bundle_list} of length 1.

An object of class \code{fhir_resource_serialized} of length 267.
}
\source{
\strong{transaction_bundle_example}\preformatted{<Bundle>
   <type value='transaction'/>
	  <entry>
	     <resource>
	        <Patient>
	           <id value='id1'/>
	           <address>
	              <use value='home'/>
	              <city value='Amsterdam'/>
	              <type value='physical'/>
	              <country value='Netherlands'/>
	           </address>
	           <name>
	              <given value='Marie'/>
	           </name>
	        </Patient>
	     </resource>
	     <request>
	        <method value='POST'/>
	        <url value='Patient'/>
	     </request>
	  </entry>
	  <entry>
	      <resource>
	         <Patient>
	            <id value='id3'/>
	            <address>
	               <use value='home'/>
	               <city value='Berlin'/>
	            </address>
	            <address>
	               <use value='work'/>
	               <city value='London'/>
	               <type value='postal'/>
	               <country value='England'/>
             </address>
	            <address>
	               <type value='postal'/>
	               <country value='France'/>
	            </address>
	            <name>
	               <given value='Frank'/>
	            </name>
	            <name>
	               <given value='Max'/>
	            </name>
	        </Patient>
	     </resource>
	     <request>
	        <method value='POST'/>
	        <url value='Patient'/>
	     </request>
	  </entry>
</Bundle>
}

\strong{transaction_bundle_example}\preformatted{<Bundle>
   <type value='transaction'/>
	  <entry>
	     <resource>
	        <Patient>
	           <address>
	              <use value='home'/>
	              <city value='Amsterdam'/>
	              <type value='physical'/>
	              <country value='Netherlands'/>
	           </address>
	           <name>
	              <given value='Marie'/>
	           </name>
	        </Patient>
	     </resource>
	     <request>
	        <method value='POST'/>
	        <url value='Patient'/>
	     </request>
	  </entry>
	  <entry>
	      <resource>
	         <Patient>
	            <address>
	               <use value='home'/>
	               <city value='Berlin'/>
	            </address>
	            <address>
	               <use value='work'/>
	               <city value='London'/>
	               <type value='postal'/>
	               <country value='England'/>
             </address>
	            <address>
	               <type value='postal'/>
	               <country value='France'/>
	            </address>
	            <name>
	               <given value='Frank'/>
	            </name>
	            <name>
	               <given value='Max'/>
	            </name>
	        </Patient>
	     </resource>
	     <request>
	        <method value='POST'/>
	        <url value='Patient'/>
	     </request>
	  </entry>
</Bundle>
}
}
\usage{
transaction_bundle_example

example_resource
}
\description{
These data examples are simple examples to try out POSTing resources
to a server. See \strong{Source} for how the xml versions look.

These data examples are simple examples to try out POSTing resources
to a server. See \strong{Source} for how the xml versions look.
}
\details{
\code{transaction_bundle_example} contains 1 transaction bundle with 2 Patient resources.

\code{example_resource} contains 1 patient resource
}
\examples{
#unserialize xml objects before doing anything else with them!
fhir_unserialize(bundles = transaction_bundle_example)
#unserialize xml objects before doing anything else with them!
fhir_unserialize(example_resource)
}
\keyword{datasets}
