% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{example_bundles1}
\alias{example_bundles1}
\alias{example_bundles2}
\title{Toy example bundles for mulitple entries}
\format{
An object of class \code{fhir_bundle_list} of length 1.

An object of class \code{fhir_bundle_list} of length 1.
}
\source{
\strong{example_bundles1}\preformatted{<Bundle>
  <Patient>
	    <id value='id1'/>
	    <address>
	       <use value='home'/>
	       <city value='Amsterdam'/>
	       <type value='physical'/>
	       <country value='Netherlands'/>
	    </address>
	    <name>
	       <given value='Marie'/>
	    </name>
	 </Patient>
	 <Patient>
	    <id value='id3'/>
	    <address>
	       <use value='home'/>
	       <city value='Berlin'/>
	    </address>
	    <address>
        <type value='postal'/>
	       <country value='France'/>
	    </address>
	    <address>
	       <use value='work'/>
	       <city value='London'/>
	       <type value='postal'/>
	       <country value='England'/>
	    </address>
	    <name>
	       <given value='Frank'/>
	    </name>
	    <name>
       <given value='Max'/>
	    </name>
  </Patient>
</Bundle>
}

\strong{example_bundles2}\preformatted{<Bundle>
	<Patient>
		<id value='id1'/>
		<address>
			<use value='home'/>
			<city value='Amsterdam'/>
			<type value='physical'/>
			<country value='Netherlands'/>
		</address>
		<name>
		<given value='Marie'/>
		</name>
	</Patient>

	<Patient>
		<id value='id2'/>
		<address>
			<use value='home'/>
			<city value='Rome'/>
			<type value='physical'/>
			<country value='Italy'/>
		</address>
		<address>
			<use value='work'/>
			<city value='Stockholm'/>
			<type value='postal'/>
			<country value='Sweden'/>
		</address>
		<name>
			<given value='Susie'/>
		</name>
	</Patient>

	<Patient>
		<id value='id3'/>
		<address>
			<use value='home'/>
			<city value='Berlin'/>
		</address>
		<address>
			<type value='postal'/>
			<country value='France'/>
		</address>
		<address>
			<use value='work'/>
			<city value='London'/>
			<type value='postal'/>
			<country value='England'/>
		</address>
		<name>
			<given value='Frank'/>
		</name>
		<name>
			<given value='Max'/>
		</name>
	</Patient>

</Bundle>
}
}
\usage{
example_bundles1

example_bundles2
}
\description{
These data examples are bundles that contain very few very simple Patient resources
that have multiple entries and can be used for demonstration purposes. See \strong{Source} for how the
xml versions look.
}
\details{
\code{example_bundles1} contains 1 bundle with 2 Patient resources.

\code{example_bundles2} contains 1 bundle with 3 Patient resources.
}
\examples{
#unserialize xml objects before doing anything else with them!
fhir_unserialize(example_bundles1)
#unserialize xml objects before doing anything else with them!
fhir_unserialize(example_bundles2)
}
\keyword{datasets}
