% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{download.all.bundles}
\alias{download.all.bundles}
\title{download.all.bundles}
\usage{
download.all.bundles(
  fhir.search.request,
  max.attempts = 5,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{fhir.search.request}{a fhir search request}

\item{max.attempts}{maximal attempts to connect to a page address}

\item{username}{a string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{a string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}
}
\value{
the downloaded bundles as a list of pages in xml format
}
\description{
downloads all fhir bunde of a fhir search request from a fhir server.
}
\examples{
\dontrun{
bundles <- download.all.bundles( "https://vonk.fire.ly/R4/Medication?_format=xml" )
}
}
