% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{download.bundle}
\alias{download.bundle}
\title{download.bundle}
\usage{
download.bundle(fhir.search.request, max.attempts = 5)
}
\arguments{
\item{fhir.search.request}{a fhir search request}

\item{max.attempts}{maximal attempts to connect to a page address}
}
\value{
the downloaded bundle as a list of pages in xml format
}
\description{
downloads a fhir bunde via fhir search request from a fhir server.
}
\examples{
\dontrun{
download.bundle( "https://hapi.fhir.org/baseR4/Medication?_format=xml" )
}
}
