% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{download.bundles}
\alias{download.bundles}
\title{download.bundles}
\usage{
download.bundles(fhir.search.request, max.attempts = 5)
}
\arguments{
\item{fhir.search.request}{a fhir search request}

\item{max.attempts}{maximal attempts to connect to a page address}
}
\value{
the downloaded bundles as a list of pages in xml format
}
\description{
downloads all fhir bunde of a fhir search request from a fhir server.
}
\examples{
\dontrun{
bundles <- download.bundles( "https://vonk.fire.ly/R4/Medication?_format=xml" )
}
}
