% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{download.page}
\alias{download.page}
\title{download.page}
\usage{
download.page(fhir.search.request, max.attempts = 5)
}
\arguments{
\item{fhir.search.request}{a fhir search request. it must contain _format=xml.}

\item{max.attempts}{the maximal number of attempts to send a request. Default is 5.}
}
\value{
the downloaded bundle page in xml format.
}
\description{
downloads a fhir bundle page via fhir search request and return it as xml file.
}
\examples{
\dontrun{
download.page( "https://hapi.fhir.org/baseR4/Medication?_format=xml" )
}
}
