% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{download.single.bundle}
\alias{download.single.bundle}
\title{download.single.bundle}
\usage{
download.single.bundle(
  fhir.search.request,
  max.attempts = 5,
  username = NULL,
  password = NULL
)
}
\arguments{
\item{fhir.search.request}{a fhir search request. it must contain _format=xml.}

\item{max.attempts}{the maximal number of attempts to send a request. Default is 5.}

\item{username}{a string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{a string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}
}
\value{
the downloaded bundle in xml format.
}
\description{
downloads a fhir bundle via fhir search request and return it as xml file.
}
\examples{
\dontrun{
download.single.bundle( "https://hapi.fhir.org/baseR4/Medication?_format=xml" )
}
}
