% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{fhir2df_auto}
\alias{fhir2df_auto}
\title{Flatten a bundles list automatically to on singe data frame.}
\usage{
fhir2df_auto(bundles, xpath = "//resource", sep = " -+- ", ns.id = NULL)
}
\arguments{
\item{bundles}{A bundles list.}

\item{xpath}{A xpath to the root node.}

\item{sep}{A separator for multiple entries.}

\item{ns.id}{A string containing the namespace id.}
}
\value{
A data frame containing the all nodes entries.


}
\description{
Flatten a bundles list automatically to on singe data frame.
}
\examples{
\dontrun{
fhir2df_auto(xml, xpath="//resource")
}
}
