% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_base}
\alias{fhir_base}
\title{Format FHIR base url}
\usage{
fhir_base(url)
}
\arguments{
\item{url}{A string containing the base URL of the FHIR server, e.g.  "http://hapi.fhir.org/baseR4"}
}
\value{
The formatted url in a named character vector
}
\description{
Takes an url string and removes leading/trailing white space and unnecessary slashes.
Is supposed to be used with \code{\link{fhir_build_url}}.
}
\examples{
fhir_base(" http://hapi.fhir.org/baseR4/")
}
