% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_bundle.R
\name{fhir_bundle_serialized}
\alias{fhir_bundle_serialized}
\title{Create \linkS4class{fhir_bundle_xml} object}
\usage{
fhir_bundle_serialized(bundle)
}
\arguments{
\item{bundle}{A serialized xml object representing a FHIR bundle}
}
\description{
Create \linkS4class{fhir_bundle_xml} object
}
\examples{
fhir_bundle_serialized(patient_bundles[[1]])
}
