% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_bundle.R
\docType{class}
\name{fhir_bundle_xml-class}
\alias{fhir_bundle_xml-class}
\title{An S4 class to represent a FHIR bundle in xml form}
\description{
A \code{fhir_bundle_xml} is an xml representation of a FHIR bundle (https://www.hl7.org/fhir/bundle.html).
It is usually found inside a \code{fhir_bundle_list} which is returned by a call to \code{\link[=fhir_search]{fhir_search()}}.
}
\section{Slots}{

\describe{
\item{\code{next_link}}{A \link[=fhir_url-class]{fhir_url} pointing to the next bundle on the server.}

\item{\code{self_link}}{A \link[=fhir_url-class]{fhir_url} pointing to this bundle on the server.}
}}

