% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_bundle.R
\name{fhir_bundle_xml}
\alias{fhir_bundle_xml}
\title{Create \link[=fhir_bundle_xml-class]{fhir_bundle_xml} object}
\usage{
fhir_bundle_xml(bundle)
}
\arguments{
\item{bundle}{A xml-object representing a FHIR bundle}
}
\description{
This should only be used if you want to create small examples. Usually, a \code{fhir_bundle_xml} will
be returned by \code{\link[=fhir_search]{fhir_search()}}.
}
\examples{
fhir_bundle_xml(bundle = xml2::xml_unserialize(patient_bundles[[1]]))

}
