% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{fhir_canonical_design}
\alias{fhir_canonical_design}
\title{Retrieve design of last call to fhir_crack}
\usage{
fhir_canonical_design()
}
\description{
Returns the \link[=fhir_design-class]{fhir_design} of the last call to \code{\link[=fhir_crack]{fhir_crack()}}.
}
\examples{
#load example bundles
bundles <- fhir_unserialize(bundles = patient_bundles)

#define design
patients <- fhir_table_description(resource = 'Patient')

design <- fhir_design(patients)

result <- fhir_crack(bundles = bundles, design = design)

fhir_canonical_design()

}
\seealso{
\code{\link[=fhir_design]{fhir_design()}}, \code{\link[=fhir_table_description]{fhir_table_description()}}
}
