% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_columns.R
\docType{class}
\name{fhir_columns-class}
\alias{fhir_columns-class}
\title{A S4 class to represent columns in a \link[=fhir_table_description-class]{fhir_table_description}}
\description{
An object of class \code{fhir_columns} is part of a \link[=fhir_table_description-class]{fhir_table_description}
in a \link[=fhir_design-class]{fhir_design} and holds information on the elements
that should be extracted from the FHIR resources, as well as the column names of the resulting data.frame.
The elements to be extracted are indicated by XPath xpaths.
}
\section{Slots}{

\describe{
\item{\code{names}}{The column names}
}}

