% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_columns.R
\name{fhir_columns}
\alias{fhir_columns}
\title{Create \linkS4class{fhir_columns} object}
\usage{
fhir_columns(expressions, colnames)
}
\arguments{
\item{expressions}{A (named) character vector or (named) list containing xpath expressions,
or a \linkS4class{fhir_xpath_expressions} object.}

\item{colnames}{The names of the columns to create. If no colnames are provided and the list or vector
in \code{expressions} has names, those names are taken as the colnames. If no colnames are provided and
\code{expressions} is unnamed too, the colnames are generated automatically from the xpath expressions. See examples.}
}
\description{
An object of class \linkS4class{fhir_columns} is part of a fhir_design and holds information on the elements
that should be extracted from the FHIR resources, as well as the column names of the resulting data.frame.
The elements to be extracted are indicated by XPath expressions. If no column names are provided,
they are generated automatically and reflect the elements position in the resource.
}
\examples{
 #provide colnames explicitly
 fhir_columns(expressions = c("name/given", "code/coding/code"),
              colnames = c("name", "code"))

 #colnames are taken from expressions argument
 fhir_columns(expressions = c(name = "name/given", code = "code/coding/code"))

 #colnames are taken from expressions argument
 fhir_columns(expressions = list(name = "name/given", code = "code/coding/code"))

 #colnames are generated automatically
 fhir_columns(expressions = c("name/given", "code/coding/code"))

}
