% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_common_columns}
\alias{fhir_common_columns}
\title{Find common columns.}
\usage{
fhir_common_columns(indexed_data_frame, column_names_prefix)
}
\arguments{
\item{indexed_data_frame}{A indexed Data Frame.}

\item{column_names_prefix}{A String containing the prefix of all common columns.}
}
\value{
A Vector of Strings reperesenting the common column names.
}
\description{
Find common columns.
}
