% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_current_search_url}
\alias{fhir_current_search_url}
\title{return FHIR search request used in last call to fhir_search}
\usage{
fhir_current_search_url()
}
\description{
return FHIR search request used in last call to fhir_search
}
