% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_design.R
\docType{class}
\name{fhir_design-class}
\alias{fhir_design-class}
\title{A S4 class containing a design for \code{\link[=fhir_crack]{fhir_crack()}}}
\description{
A fhir_design is a named list of \link[=fhir_table_description-class]{fhir_table_description} objects. Each table_description
contains information on how to flatten one resource type which will result in one
data.frame. The fhir_design is passed to the function \code{\link[=fhir_crack]{fhir_crack()}} along with a
list of bundles containing FHIR resources.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{The list of \code{fhir_table_description} objects.}

\item{\code{names}}{The names of the table_descriptions. Those will also be the names of the
resulting data.frames.}
}}

\seealso{
\code{\link[=fhir_table_description]{fhir_table_description()}}, \code{\link[=fhir_crack]{fhir_crack()}}
}
