% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_df_description.R
\name{fhir_df_description}
\alias{fhir_df_description}
\title{Create \linkS4class{fhir_df_description} object}
\usage{
fhir_df_description(resource, cols, style = fhir_style())
}
\arguments{
\item{resource}{A character vector of length one or \linkS4class{fhir_resource_type} object,
indicating which resource type should be extracted.}

\item{cols}{A \linkS4class{fhir_columns} object or something that can be coerced to one,
like a(named) character vector, a (named) list containing xpath expressions,
or a \linkS4class{fhir_xpath_expressions} object. See \code{\link[=fhir_columns]{fhir_columns()}} and the examples.}

\item{style}{Optional A \linkS4class{fhir_style} object, as created by \code{\link[=fhir_style]{fhir_style()}}.
If this argument is omitted, default values will be assumed, see \code{\link[=fhir_style]{fhir_style()}}.}
}
\value{
An object of class \linkS4class{fhir_df_description}.
}
\description{
Create \linkS4class{fhir_df_description} object
}
\examples{
#named character for cols
fhir_df_description(resource = "Patient",
                    cols = c(name = "name/family",
                             gender = "gender",
                             id = "id"),
                    style = fhir_style(sep = "||",
                                       brackets = c("[", "]"),
                                       rm_empty_cols = FALSE
                    )
)

#' #If style is ommitted, default values are assumed
fhir_df_description(resource = "Patient",
                    cols = c(name = "name/family",
                             gender = "gender",
                             id = "id")
)

#named list for cols
fhir_df_description(resource = "Patient",
                    cols = list(name = "name/family",
                             gender = "gender",
                             id = "id")
)

#unnamed character for cols, colnames are generated automatically
fhir_df_description(resource = "Patient",
                    cols = c("name/family",
                                "gender",
                                "id")
)

}
