% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_table_list.R
\docType{class}
\name{fhir_dt_list-class}
\alias{fhir_dt_list-class}
\title{List of data.tables as returned by \code{\link[=fhir_crack]{fhir_crack()}}}
\description{
Objects of this class are returned by \code{\link[=fhir_crack]{fhir_crack()}} when \code{data.table=TRUE}.
They behave like an ordinary named list of data.tables but have some additional information
in the slot \code{design}.
}
\section{Slots}{

\describe{
\item{\code{names}}{A character vector containing the names of the data.tables.}

\item{\code{design}}{An object of class \link[=fhir_design-class]{fhir_design} that was used to create the dt_list.}
}}

