% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_extract}
\alias{fhir_extract}
\title{Extracting Multiple Entries}
\usage{
fhir_extract(
  indexed_data_frame,
  column.prefix = "id",
  brackets = c("<", ">"),
  sep = " -+- ",
  id_name = "RESOURCE_IDENTIFICATOR",
  all_columns = F
)
}
\arguments{
\item{indexed_data_frame}{A Data Frame with indexed multiple entries in its columns.}

\item{column.prefix}{A String specifiying a common prefix of the names of all columns to be extracted simultaneously.}

\item{brackets}{A character vector of Length 2, holding the Brackets.}

\item{sep}{A string, the separator.}

\item{id_name}{A String, the name of the column holding the resource id (row id).}

\item{all_columns}{A Logical Scalar. Return all columns or extracted only.}
}
\value{
A data frame without indices
}
\description{
Extracting Multiple Entries
}
\examples{
\dontrun{
fhir_extrac( df )
}

}
