% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_key_value}
\alias{fhir_key_value}
\title{Build key value pairs for FHIR search}
\usage{
fhir_key_value(key, value, url_enc = TRUE)
}
\arguments{
\item{key}{The name of the search parameter, e.g. "_include", "gender" or "_summary".
For a general overview see https://www.hl7.org/fhir/search.html
and also check out the paragraph on search parameters for the respective ressource,
e.g. http://www.hl7.org/fhir/patient.html#search}

\item{value}{The name of the respective value for the parameter, e.g. "Observation:patient",
"female" or "count".}

\item{url_enc}{URL encode key value pairs? Defaults to TRUE, which is advisable in most cases.}
}
\value{
A string with the appropriately encoded key value pairs
}
\description{
Takes two strings representing a key value pair for a FHIR search parameter and
encodes the pair properly.  \code{fhir_key_value} is supposed to be used with
\code{\link{fhir_build_search_url}}
}
\examples{
fhir_key_value(key = "gender", value = "female")
fhir_key_value(key = "category", value = "http://snomed.info/sct|116223007")
}
