% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_key_value_pair.R
\docType{class}
\name{fhir_key_value_pair-class}
\alias{fhir_key_value_pair-class}
\title{An S4 class to represent a key value pair of a FHIR search parameter
#'}
\description{
An S4 class to represent a key value pair of a FHIR search parameter
#'
}
\section{Slots}{

\describe{
\item{\code{keys}}{A length 1 character defining the parameter key}

\item{\code{values}}{A length 1 character defining the parameter value}
}}

