% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_key_value_pair.R
\name{fhir_key_value_pair}
\alias{fhir_key_value_pair}
\title{Create \linkS4class{fhir_key_value_pair} object}
\usage{
fhir_key_value_pair(key, value)
}
\arguments{
\item{key}{A length 1 character specifying a search parameter key, e.g \code{"gender"} or \code{"_summary"}}

\item{value}{A length 1 character specifying a search parameter value, e.g \code{"male"} or \code{"count"}}
}
\description{
This function creates an URL encoded key value pair for FHIR search parameters
}
\examples{
fhir_key_value_pair(key="gender", value="male")
fhir_key_value_pair(key="_summary", value="count")
}
