% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{fhir_load_design}
\alias{fhir_load_design}
\title{Load design from xml}
\usage{
fhir_load_design(file)
}
\arguments{
\item{file}{A string specifying the file from which to read.}
}
\value{
A \link[=fhir_design-class]{fhir_design} object. See \code{?fhir_design}.
}
\description{
Loads a \link[=fhir_design-class]{fhir_design} for use with \code{\link[=fhir_crack]{fhir_crack()}} from an xml file into R.
}
\examples{

table_desc1 <- fhir_table_description(
    resource = 'Patient',
    cols     = c(
        id     = 'id',
        name   = 'name/family',
        gender = 'gender'
    ),
    sep           = ':::',
    brackets      = c('[', ']'),
    rm_empty_cols = FALSE,
    format        = 'compact',
    keep_attr     = FALSE
)

table_desc2 <- fhir_table_description(
    resource = 'Observation',
    cols     = c(
        'code/coding/system',
        'code/coding/code'
    )
)

design <- fhir_design(
    Patients     = table_desc1,
    Observations = table_desc2
)
temp <- tempfile()

fhir_save_design(design = design, file = temp)

design <- fhir_load_design(file = temp)

}
\seealso{
\code{\link[=fhir_design]{fhir_design()}}, \code{\link[=fhir_table_description]{fhir_table_description()}}, \code{\link[=fhir_save_design]{fhir_save_design()}}
}
