% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_resources.R
\docType{methods}
\name{fhir_post}
\alias{fhir_post}
\alias{fhir_post,ANY,fhir_resource-method}
\alias{fhir_post,fhir_resource-method}
\alias{fhir_post,ANY,fhir_bundle_xml-method}
\alias{fhir_post,fhir_bundle_xml-method}
\alias{fhir_post,ANY,fhir_body-method}
\alias{fhir_post,fhir_body-method}
\title{POST to a FHIR server}
\usage{
fhir_post(
  url,
  body,
  username = NULL,
  password = NULL,
  token = NULL,
  add_headers = NULL,
  verbose = 1,
  log_errors = NULL
)

\S4method{fhir_post}{ANY,fhir_resource}(
  url,
  body,
  username = NULL,
  password = NULL,
  token = NULL,
  add_headers = NULL,
  verbose = 1,
  log_errors = NULL
)

\S4method{fhir_post}{ANY,fhir_bundle_xml}(
  url,
  body,
  username = NULL,
  password = NULL,
  token = NULL,
  add_headers = NULL,
  verbose = 1,
  log_errors = NULL
)

\S4method{fhir_post}{ANY,fhir_body}(
  url,
  body,
  username = NULL,
  password = NULL,
  token = NULL,
  add_headers = NULL,
  verbose = 1,
  log_errors = NULL
)
}
\arguments{
\item{url}{An object of class \link[=fhir_url-class]{fhir_url} or a character vector of length one containing the url to POST to.}

\item{body}{An object of class  \link[=fhir_resource-class]{fhir_resource}, \link[=fhir_bundle_xml-class]{fhir_bundle_xml} or \link[=fhir_body-class]{fhir_body}.
See details for how to generate them.}

\item{username}{A character vector of length one containing the username for basic authentication.}

\item{password}{A character vector of length one containing the password for basic authentication.}

\item{token}{A character vector of length one or object of class \link[httr:Token-class]{httr::Token}, for bearer token authentication (e.g. OAuth2). See \code{\link[=fhir_authenticate]{fhir_authenticate()}}
for how to create this.}

\item{add_headers}{A named character vector of custom headers to add to the HTTP request, e.g. \code{c(myHeader = "somevalue")} or
\code{c(firstHeader = "value1", secondHeader = "value2")}.}

\item{verbose}{An integer vector of length one. If 0, nothing is printed, if > 0 success message is printed. Defaults to 1.}

\item{log_errors}{Either \code{NULL} or a character vector of length one indicating the name of a file in which to save http errors.
\code{NULL} means no error logging. When a file name is provided, the errors are saved in the specified file. Defaults to \code{NULL}.
Regardless of the value of \code{log_errors} the most recent http error message within the current R session is saved internally and can
be accessed with \code{\link[=fhir_recent_http_error]{fhir_recent_http_error()}}.}
}
\description{
This function is a convenience wrapper around \code{\link[httr:POST]{httr::POST()}}.
}
\details{
\code{\link[=fhir_post]{fhir_post()}} accepts four classes for the body:
\enumerate{
\item A \link[=fhir_resource-class]{fhir_resource} as created by \code{\link[=fhir_build_resource]{fhir_build_resource()}}. This is used when just a single resource should be POSTed to the server.
In this case \code{url} must contain the base url plus the resource type, e.g. http://hapi.fhir.org/baseR4/Patient.
\item A \link[=fhir_bundle_xml-class]{fhir_bundle_xml} representing a transaction or batch bundle as created by \code{\link[=fhir_build_bundle]{fhir_build_bundle()}}.
\item A \link[=fhir_body-class]{fhir_body} as created by \code{\link[=fhir_body]{fhir_body()}}. This is the most flexible approach, because within the \link[=fhir_body-class]{fhir_body} object you can represent
any kind of \code{content} as a string and set the \code{type} accordingly. See examples.
}

For examples of how to create the different body types see the respective help pages. For an example of the entire workflow around creating
and POSTing resources, see the package vignette on recreating resources.
}
\examples{
\dontrun{
### 1. POST transaction bundle
#unserialize example bundles
bundle <- fhir_unserialize(transaction_bundle_example)

#have a look at the bundle
cat(toString(bundle))

#post
fhir_post(url = "http://hapi.fhir.org/baseR4", body = bundle)


### 2. POST single resouce
#unserialize example resource
resource <- fhir_unserialize(example_resource1)

#have a look at the resource
resource

#post
url <- fhir_url(url = "http://hapi.fhir.org/baseR4", resource = "Patient")
fhir_post(url = url, body = resource)


### 3. POST arbitrary body
#define body
body <- fhir_body(content = "<Patient> <gender value='female'/> </Patient>", type = "xml")

#post
url <- fhir_url(url = "http://hapi.fhir.org/baseR4", resource = "Patient")
fhir_post(url = url, body = body)
}
}
