% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{fhir_remove_div}
\alias{fhir_remove_div}
\title{Remove <div>-Blocks from a String}
\usage{
fhir_remove_div(string)
}
\arguments{
\item{string}{A character vector containing xml.}
}
\value{
A character vector containing xml without <div>-Blocks.
}
\description{
Remove <div>-Blocks from a String
}
\examples{
fhir_remove_div(c("Hallo<div><p>Remove Me</p></div> World!", "A<div><div>B</div>C</div>D"))
}
