% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{fhir_remove_div_from_bundle}
\alias{fhir_remove_div_from_bundle}
\title{Remove <div>-Blocks from a XML-Bundle}
\usage{
fhir_remove_div_from_bundle(bundle)
}
\arguments{
\item{bundle}{A fhir_bundle_xml object.}
}
\value{
A fhir_bundle_xml object without <div>-Blocks
}
\description{
Remove <div>-Blocks from a XML-Bundle
}
\examples{
b1 <- fhir_unserialize(patient_bundles)
b2 <- fhir_bundle_list(lapply(b1, fhir_remove_div_from_bundle))
(sb1 <- sum(sapply(b1, function(x)object.size(toString(x)))))
(sb2 <- sum(sapply(b2, function(x)object.size(toString(x)))))
#div makes about 29 percent of patient bundles
100 * (sb1 - sb2) / sb1
cat(toString(b1[[1]]))
cat(toString(b2[[1]]))
}
