% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{fhir_remove_tag}
\alias{fhir_remove_tag}
\title{Remove a certain Tag from a String}
\usage{
fhir_remove_tag(string, tag = "div")
}
\arguments{
\item{string}{A character vector containing an xml-strings.}

\item{tag}{A character of length 1 containing the tag tha should be removed.}
}
\value{
A character vector containing the xml without the tag.
}
\description{
Remove a certain Tag from a String
}
\examples{
fhir_remove_tag(c("Hallo<div><p>Remove Me</p></div> World!", "A<div><div><p>B</p></div>C</div>D"), "p")
}
