% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{fhir_remove_tag_from_bundle}
\alias{fhir_remove_tag_from_bundle}
\title{Remove a certain Tag from a XML-Bundle}
\usage{
fhir_remove_tag_from_bundle(bundle, tag = "div")
}
\arguments{
\item{bundle}{A fhir_bundle_xml object.}

\item{tag}{A character of length 1 containing the tag that should be removed.}
}
\value{
A fhir_bundle_xml containing xmls without the tag.
}
\description{
Remove a certain Tag from a XML-Bundle
}
\examples{
b1 <- fhir_unserialize(patient_bundles)
b2 <- fhir_bundle_list(lapply(b1, fhir_remove_tag_from_bundle, tag = "p"))
(sb1 <- sum(sapply(b1, function(x)object.size(toString(x)))))
(sb2 <- sum(sapply(b2, function(x)object.size(toString(x)))))
#div makes about 29 percent of patient bundles
100 * (sb1 - sb2) / sb1
cat(toString(b1[[1]]))
cat(toString(b2[[1]]))
}
