% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_tag.R
\name{fhir_remove_tag_from_bundle_list}
\alias{fhir_remove_tag_from_bundle_list}
\title{Remove a certain Tag from a XML-Bundle}
\usage{
fhir_remove_tag_from_bundle_list(bundle_list, tag = "div")
}
\arguments{
\item{bundle_list}{A fhir_bundle_list object containing xml-bundles from where the Tag should be removed.}

\item{tag}{A character of length 1 containing the Tag to remove.}
}
\value{
A fhir_bundle_list object containing xml-bundles from where the Tag is removed.
}
\description{
Remove a certain Tag from a XML-Bundle
}
\examples{
b1 <- fhir_unserialize(patient_bundles)
b2 <- fhir_remove_tag_from_bundle_list(b1)
(sb1 <- sum(sapply(b1, function(x)object.size(toString(x)))))
(sb2 <- sum(sapply(b2, function(x)object.size(toString(x)))))
#div makes about 29 percent of patient bundles
100 * (sb1 - sb2) / sb1
cat(toString(b1[[1]]))
cat(toString(b2[[1]]))
}
