% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R, R/fhir_resource.R
\name{fhir_resource}
\alias{fhir_resource}
\title{Check and format FHIR resource type for FHIR search}
\usage{
fhir_resource(string)

fhir_resource(string)
}
\arguments{
\item{string}{A string containing the resource type. Will mostly be one of the official FHIR resource
types listed at https://hl7.org/FHIR/resourcelist.html}

\item{resource}{A string containing the resource type for the FHIR search.
Should be one of the official FHIR resource types listed at https://hl7.org/FHIR/resourcelist.html}
}
\value{
A named character vector with the checked and formatted resource type

An fhir_resource object
}
\description{
This function takes a string defining a FHIR resource type and formats it correctly,
removing white space and slashes. It also checks the resource against the list
of resources provided at https://hl7.org/FHIR/resourcelist.html and throws a warning
if the resource doesn't match. \code{fhir_resource} is supposed to be used with
\code{\link{fhir_build_request}}.

This function creates an object of class \code{fhir_resource} by taking a string defining a FHIR resource type
and formating it correctly,i.e. removing white space and slashes. It also checks the resource against the list
of resources provided at https://hl7.org/FHIR/resourcelist.html and converts cases appropriately
if the resource is found in that list.
}
\examples{
fhir_resource("patient")
fhir_resource("MedicationAdministration")
}
