% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_resource_type.R
\docType{class}
\name{fhir_resource_type-class}
\alias{fhir_resource_type-class}
\title{A representation of a FHIR resource type}
\description{
An object of class \code{fhir_resource_type} is a string containing a FHIR resource type.
It is part of a \code{fhir_table_description} which in turn is part of a \code{fhir_design} and used in
\code{\link[=fhir_crack]{fhir_crack()}}.
}
