% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_resource.R
\docType{class}
\name{fhir_resource_xml-class}
\alias{fhir_resource_xml-class}
\title{An S4 class to represent a FHIR resource in xml form}
\description{
A \code{fhir_resource_xml} is an xml representation of a FHIR resource (https://www.hl7.org/fhir/resourcelist.html).
}
