% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_restore_indices}
\alias{fhir_restore_indices}
\title{Restore indices to molten data.frame with multiple entries}
\usage{
fhir_restore_indices(d, index_matrix)
}
\arguments{
\item{d}{A data.frame/data.table as produced by \code{\link{fhir_melt_all}} without indices.}

\item{index_matrix}{A character matrix with the same dimensions as \code{d}, as produced by \code{\link{fhir_extract_indices}}}
}
\value{
\code{d} but with the indices from \code{index_matrix}
}
\description{
Takes a molten data.frame/data.table containing no indices as produced by \code{\link{fhir_melt_all}} and
restores its indices from a matrix of indices as produces by \code{\link{fhir_extract_indices}}. Dimensions as
well as row and column ordering of index matrix and data frame have to be identical!
}
\examples{
#generate example bundle
bundle <- xml2::read_xml(
"<Bundle>

	<Patient>
		<id value='id1'/>
		<address>
			<use value='home'/>
			<city value='Amsterdam'/>
			<type value='physical'/>
			<country value='Netherlands'/>
		</address>
		<birthDate value='1992-02-06'/>
	</Patient>

	<Patient>
		<id value='id2'/>
		<address>
			<use value='home'/>
			<city value='Rome'/>
			<type value='physical'/>
			<country value='Italy'/>
		</address>
		<address>
			<use value='work'/>
			<city value='Stockholm'/>
			<type value='postal'/>
			<country value='Sweden'/>
		</address>
		<birthDate value='1980-05-23'/>
	</Patient>

	<Patient>
		<id value='id3'/>
		<address>
			<use value='home'/>
			<city value='Berlin'/>
		</address>
		<address>
			<type value='postal'/>
			<country value='France'/>
		</address>
		<address>
			<use value='work'/>
			<city value='London'/>
			<type value='postal'/>
			<country value='England'/>
		</address>
		<birthDate value='1974-12-25'/>
		<birthDate value='1978-11-13'/>
	</Patient>

</Bundle>"
)

#crack fhir resources
dfs <- fhir_crack(bundles = list(bundle),
                 design = list(Patients = list(resource = ".//Patient",
			                                  style= list(brackets = c("[","]"),
			                                   sep="||"))))
#Melt multiple entries
d <- fhir_melt_all(dfs$Patients, brackets = c("[", "]"), sep="||", rm_indices = F)

#Extract indices
indices <- fhir_extract_indices(d, brackets = c("[", "]"))

#remove indices
d_removed <- fhir_rm_indices(d, brackets = c("[", "]"))

#restore indices
d_restored <- fhir_restore_indices(d_removed, indices)

#compare
identical(d, d_restored)
}
