% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_url.R
\docType{class}
\name{fhir_search_url-class}
\alias{fhir_search_url-class}
\title{An S4 class to represent a URL for a FHIR search request}
\description{
Objects of this class should always be created with a call to \code{\link[=fhir_search_url]{fhir_search_url()}}
}
\section{Slots}{

\describe{
\item{\code{base}}{A character vector of length 1 containing the base url to the FHIR server}

\item{\code{resource}}{An object of type \linkS4class{fhir_resource_type} defining the resource type that is searched}

\item{\code{parameters}}{An object of type \linkS4class{fhir_parameters} defining search parameters}
}}

