% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_table_list.R
\docType{class}
\name{fhir_table_list-class}
\alias{fhir_table_list-class}
\title{Virtual superclass for \linkS4class{fhir_df_list} and \linkS4class{fhir_dt_list}}
\description{
Virtual superclass for \linkS4class{fhir_df_list} and \linkS4class{fhir_dt_list}
}
\section{Slots}{

\describe{
\item{\code{names}}{Character vector containing the names of the tables}

\item{\code{indexed}}{Logical vector indicating which of the tables has indices for multiple entries}

\item{\code{design}}{An object of class \linkS4class{fhir_design} that was used to create the table list}
}}

