% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_tree.R
\name{fhir_tree.apply}
\alias{fhir_tree.apply}
\title{Apply Functions on a Tree}
\usage{
fhir_tree.apply(tree, fun.start = NULL, fun.finish = NULL)
}
\arguments{
\item{tree}{A tree as build by fhir_tree.new() from a wide cracked table.}

\item{fun.start}{A Function called before going deeper into the tree.}

\item{fun.finish}{A Function called after going deeper into the tree.}
}
\value{
A Tree.
}
\description{
Apply Functions on a Tree
}
\examples{
#unserialize example
bundles <- fhir_unserialize(bundles = example_bundles1)

#crack fhir resources
table_desc <- fhir_table_description(
    resource = "Patient",
    brackets = c("[", "]"),
    sep      = " "
)
df <- fhir_crack(bundles = bundles, design = table_desc)

#cast silently using bracktets and separator definitions from table_desc
cast_df <- fhir_cast(df, brackets = table_desc@brackets, sep = table_desc@sep, verbose = 0)

#build tree
tree <- fhir_tree.new(cast_df, brackets = table_desc@brackets, root = "Patient")
tree_applied <- fhir_tree.apply(tree, fhir_tree.fun.rm_ids)
tree_faster <- fhir_tree.rm_ids(tree)
cat(fhir_tree.as_text(tree_applied))
cat(fhir_tree.as_text(tree_faster))
}
