% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_tree.R
\name{fhir_tree.get_attr}
\alias{fhir_tree.get_attr}
\title{Get Attributes of a Tree Node}
\usage{
fhir_tree.get_attr(tree)
}
\arguments{
\item{tree}{A tree as build by fhir_tree.new}
}
\value{
A list of attributes' names and values
}
\description{
Get Attributes of a Tree Node
}
