% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_tree.R
\name{fhir_tree.new}
\alias{fhir_tree.new}
\title{Build Tree from Table}
\usage{
fhir_tree.new(table, brackets, root)
}
\arguments{
\item{table}{A wide cracked table.}

\item{brackets}{A character of length one or two containing the brackets used in the column names of the table.}

\item{root}{A character of length one, the resource name.}

\item{skip_one}{A logical of length one indicating whether index 1 is skipped or not.}
}
\value{
A tree.
}
\description{
Build Tree from Table
}
\examples{
#unserialize example
bundles <- fhir_unserialize(bundles = example_bundles1)

#crack fhir resources
table_desc <- fhir_table_description(
    resource = "Patient",
    brackets = c("[", "]"),
    sep      = " "
)
df <- fhir_crack(bundles = bundles, design = table_desc)

#cast silently using bracktets and separator definitions from table_desc
cast_df <- fhir_cast(df, brackets = table_desc@brackets, sep = table_desc@sep, verbose = 0)

#build tree
tree <- fhir_tree.new(cast_df, brackets = table_desc@brackets, root = "Patient")
cat(fhir_tree.as_text(tree))
}
