% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_tree.R
\name{fhir_tree.rm_ids}
\alias{fhir_tree.rm_ids}
\title{Remove ids from tree
Removes the ids leftover from the casted table}
\usage{
fhir_tree.rm_ids(tree)
}
\arguments{
\item{tree}{A tree as produced by \code{\link[=fhir_tree.new]{fhir_tree.new()}}}
}
\description{
Remove ids from tree
Removes the ids leftover from the casted table
}
\examples{
#unserialize example
bundles <- fhir_unserialize(bundles = example_bundles1)

#crack fhir resources
table_desc <- fhir_table_description(
    resource = "Patient",
    brackets = c("[", "]"),
    sep      = " "
)

df <- fhir_crack(bundles = bundles, design = table_desc)

#cast
cast_df <- fhir_cast(df, brackets = table_desc@brackets, sep = table_desc@sep, verbose = 0)

#build tree
tree <- fhir_tree.new(cast_df, brackets = table_desc@brackets, root = "Patient")

tree <- fhir_tree.rm_ids(tree)
cat(fhir_tree.as_text(tree))
}
