% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R
\docType{methods}
\name{fhir_unserialize}
\alias{fhir_unserialize}
\alias{fhir_unserialize,fhir_bundle_xml-method}
\alias{fhir_unserialize,fhir_bundle_serialized-method}
\alias{fhir_unserialize,fhir_resource_xml-method}
\alias{fhir_unserialize,fhir_resource_serialized-method}
\alias{fhir_unserialize,fhir_bundle_list-method}
\title{Unserialize a \link[=fhir_bundle-class]{fhir_bundle}, \link[=fhir_bundle_list-class]{fhir_bundle_list} or \link[=fhir_resource-class]{fhir_resource}}
\usage{
fhir_unserialize(bundles)

\S4method{fhir_unserialize}{fhir_bundle_xml}(bundles)

\S4method{fhir_unserialize}{fhir_bundle_serialized}(bundles)

\S4method{fhir_unserialize}{fhir_resource_xml}(bundles)

\S4method{fhir_unserialize}{fhir_resource_serialized}(bundles)

\S4method{fhir_unserialize}{fhir_bundle_list}(bundles)
}
\arguments{
\item{bundles}{A \link[=fhir_bundle-class]{fhir_bundle}, \link[=fhir_bundle_list-class]{fhir_bundle_list} or \link[=fhir_resource-class]{fhir_resource} object.}
}
\value{
A  \link[=fhir_bundle_serialized-class]{fhir_bundle_serialized}, \link[=fhir_bundle_list-class]{fhir_bundle_list} or \link[=fhir_resource_serialized-class]{fhir_resource_serialized}object.
}
\description{
Unserializes FHIR resources or bundles that have been serialized to allow for saving in .rda or .RData format,
i.e. it turns a \link[=fhir_bundle_serialized-class]{fhir_bundle_serialized}/\link[=fhir_resource_serialized-class]{fhir_resource_serialized} object into an \link[=fhir_bundle_xml-class]{fhir_bundle_xml}/\link[=fhir_resource_xml-class]{fhir_resource_xml} object.
}
\examples{

#unserialize bundle list
fhir_unserialize(patient_bundles)

#unserialize single bundle
fhir_unserialize(patient_bundles[[1]])

}
