% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_unserialize}
\alias{fhir_unserialize}
\title{Serialize a FHIR Bundle list}
\usage{
fhir_unserialize(bundles)
}
\arguments{
\item{bundles}{A list of xml objects representing FHIR bundles as returned by \code{\link{fhir_search}}}
}
\value{
A list of serialized xml objects
}
\description{
Unserializes a list of FHIR bundles that have been serialized to allow for saving in .rda or .RData format.
}
\examples{
bundles <- fhir_unserialize(medication_bundles)
}
