% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_update_search_url}
\alias{fhir_update_search_url}
\title{Update the URL from the last call to \code{\link{fhir_search}}}
\usage{
fhir_update_search_url(..., append = FALSE, return_url = TRUE)
}
\arguments{
\item{...}{calls to \code{\link{fhir_key_value}}}

\item{append}{Logical. Keep key value pairs from previous request? Defaults to FALSE,
meaning only base url and resource type from previous request are kept.}

\item{return_url}{Logical. Return string with updated URL? Defaults to TRUE.}
}
\value{
A string with the updated fhir search request or NULL.
}
\description{
Update the URL from the last call to \code{\link{fhir_search}}
}
