% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_xpath_expression.R
\docType{class}
\name{fhir_xpath_expression-class}
\alias{fhir_xpath_expression-class}
\title{An S4 class for xpath_expressions
Objects of this class are essentially character vectors, but can only be valid
XPath (1.0) expressions.
They are mostly used in the \code{fhir_columns} class.}
\description{
An S4 class for xpath_expressions
Objects of this class are essentially character vectors, but can only be valid
XPath (1.0) expressions.
They are mostly used in the \code{fhir_columns} class.
}
