% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{get.bundle}
\alias{get.bundle}
\title{get_bundle}
\usage{
get.bundle(
  request,
  username = NULL,
  password = NULL,
  verbose = T,
  max.attempts = 5,
  delay.between.attempts = 10
)
}
\arguments{
\item{request}{A string containing the full fhir search request.}

\item{username}{A string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{A string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}

\item{verbose}{A logical scalar. Should downloading information be printed to the console? Defaults to TRUE.}

\item{max.attempts}{A numeric scalar. The maximal number of attempts to send a request, defaults to 5.}

\item{delay.between.attempts}{A scalar numeric specifying the delay in seconds between two attempts. Defaults to 10.}
}
\value{
The downloaded bundle in xml format.
}
\description{
Downloads a single fhir bundle via fhir search request and return it as a xml object.
}
\examples{
\dontrun{
get.bundle( request = "https://hapi.fhir.org/baseR4/Medication?_count=500&_format=xml" )
}
}
