% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{get.bundles}
\alias{get.bundles}
\title{get.bundles}
\usage{
get.bundles(
  request,
  username = NULL,
  password = NULL,
  max.bundles = Inf,
  verbose = T,
  max.attempts = 5,
  delay.between.attempts = 10
)
}
\arguments{
\item{request}{a fhir search request}

\item{username}{a string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{a string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}

\item{max.bundles}{maximal number of bundles to get. Defaults to Inf meaning all available bundles are downloaded.}

\item{verbose}{print downloading progress to console? Defaults to TRUE.}

\item{max.attempts}{maximal attempts to connect to a page address. Default is 5.}

\item{delay.between.attempts}{a delay in seconds between two attempts.}
}
\value{
the downloaded bundles as a list of pages in xml format
}
\description{
downloads all fhir bunde of a fhir search request from a fhir server.
}
\examples{
\dontrun{
bundles <- get.bundles( "https://vonk.fire.ly/R4/Medication?_format=xml" )
}
}
