% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_bundle}
\alias{get_bundle}
\title{Download single FHIR bundle}
\usage{
get_bundle(
  request,
  username = NULL,
  password = NULL,
  verbose = T,
  max_attempts = 5,
  delay_between_attempts = 10
)
}
\arguments{
\item{request}{A string containing the full FHIR search request.}

\item{username}{A string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{A string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}

\item{verbose}{A logical scalar. Should downloading information be printed to the console? Defaults to TRUE.}

\item{max_attempts}{A numeric scalar. The maximal number of attempts to send a request, defaults to 5.}

\item{delay_between_attempts}{A numeric scalar specifying the delay in seconds between two attempts. Defaults to 10.}
}
\value{
The downloaded bundle in xml format.
}
\description{
Download a single FHIR bundle via FHIR search request and return it as a xml object.
}
\examples{
bundle<-fhircrackr:::get_bundle(request = "https://hapi.fhir.org/baseR4/Patient?")
}
