% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{get_resources_by_ids}
\alias{get_resources_by_ids}
\title{Get Resources by their IDs}
\usage{
get_resources_by_ids(endpoint, resource, ids, verbose = 0)
}
\arguments{
\item{endpoint}{A character of length 1 containing the FHIR R4 endpoint.}

\item{resource}{A character of length 1 containing the Resource's name.}

\item{ids}{A character containing the IDs.}

\item{verbose}{An integer of length 1 containing the level of verbosity. Defaults to 0.}
}
\value{
A list of bundles containing the downloaded resources.
}
\description{
Get Resources by their IDs
}
\examples{
ids <- get_resources_ids(
  endpoint   = 'https://hapi.fhir.org/baseR4',
  resource   = "Patient",
  parameters = "name=Homer&name=Simpson")
ids <- sample(ids, 10)
bndls <- get_resources_by_ids(
  endpoint = 'https://hapi.fhir.org/baseR4',
  resource = "Patient",
  ids      = ids)
fhir_crack(
  bndls,
  fhir_table_description(
    resource = "Patient",
    cols     = list(
      ID      = "id",
      Homer   = "name/given",
      Simpson = "name/family")))
}
