% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{get_resources_count}
\alias{get_resources_count}
\title{Get Resources' Counts}
\usage{
get_resources_count(endpoint, resource, parameters = NULL, verbose = 0)
}
\arguments{
\item{endpoint}{A character of length 1 containing the FHIR R4 endpoint.}

\item{resource}{A character of length 1 containing the Resource's name.}

\item{parameters}{Either a character of length 1 containing the parameter string or
a named vector of the form c("gender" = "male", "_summary" = "count").}

\item{verbose}{An integer of length 1 containing the level of verbosity. Defaults to 0.}
}
\value{
An integer of length 1 containing the number of resources named \code{resource}
filtered by parameters available on the FHIR R4 endpoint.
}
\description{
Get Resources' Counts
}
\examples{
get_resources_count(
  endpoint = 'https://hapi.fhir.org/baseR4',
  resource = "Patient",
  parameters = c(gender = "female", name = "Simpson"))
}
