% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_invalid_design}
\alias{is_invalid_design}
\title{Check design}
\usage{
is_invalid_design(design)
}
\arguments{
\item{design}{The design to be checked}
}
\value{
TRUE if design is invalid, FALSE if design is valid
}
\description{
Checks whether a design provided to \code{\link{fhir_crack}} is valid and
issues a warning if it is not.
}
