% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{load_bundles}
\alias{load_bundles}
\title{Load bundles from xml-files}
\usage{
load_bundles(directory)
}
\arguments{
\item{directory}{A string containing the path to the folder were the data are stored.}
}
\value{
A list of bundles in xml format.
}
\description{
Reads all bundles stored as xml files from a directory.
}
\examples{
\dontrun{
bundles.bak <- load_bundles("result")
}
}
