% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{lst}
\alias{lst}
\title{Transform vector to named list}
\usage{
lst(..., prefix = NULL, suffix = NULL)
}
\arguments{
\item{...}{A vector that can be coerced to a character.}

\item{prefix}{A string taken as the prefix for the names of the list elements.}

\item{suffix}{A string taken as the suffix for the names of the list elements.}
}
\value{
A named list, where the names are the content surrounded by a prefix and a suffix.
}
\description{
Transforms a vector of items to a named list. The names are created with a prefix and a suffix surrounding the items.
}
\examples{
lst( LETTERS[ 1 : 5 ], prefix = "id[", suffix = "]" )
}
