% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{medication_bundles}
\alias{medication_bundles}
\title{Exemplary FHIR bundles}
\format{
\linkS4class{fhir_bundle_list} containing 1 \linkS4class{fhir_bundle_serialized} object.
\emph{They have to be unserialized before use, see Examples!}
}
\source{
Generated by

\code{
search_request  <- fhir_url("https://hapi.fhir.org/baseR4",
							resource = "MedicationStatement",
							parameters = c("code" = "http://snomed.info/ct|429374003",
										   "_include" = "MedicationStatement:subject"))

bundles <- fhir_search(search_request, max_bundles = 1)

medication_bundles <- fhir_serialize(bundles)
}

\strong{Downloaded 04-26-21}
}
\usage{
medication_bundles
}
\description{
This data example can be used to explore some of the functions from the fhircrackr package
when direct access to a FHIR server is not possible.
}
\details{
\code{medication_bundles} is a \linkS4class{fhir_bundle_list} containing \linkS4class{fhir_bundle_serialized} objects
representing FHIR bundles as returned by \code{\link[=fhir_search]{fhir_search()}}.

It contains 1 bundle with MedicationStatement resources representing Medications with Snomed CT code
429374003 and the respective Patient resources that are linked to these MedicationStatements.

It corresponds to the example of downloading and flattening FHIR resources from the vignette of the package.
}
\examples{
#unserialize xml objects before doing anything else with them!
medication_bundles <- fhir_unserialize(medication_bundles)


}
\keyword{datasets}
