% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{order_strings_with_numbers_correctly}
\alias{order_strings_with_numbers_correctly}
\title{Order Strings with Numbers correctly}
\usage{
order_strings_with_numbers_correctly(s)
}
\arguments{
\item{s}{A character containing the strings to be ordered.}
}
\value{
A character containing the order of the given strings.
}
\description{
Order Strings with Numbers correctly
}
\examples{
s <- c("12", "2", "1", "21")
order(s)
order_strings_with_numbers_correctly(s = s)
}
