% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{page.to.dataframes}
\alias{page.to.dataframes}
\title{page.to.dataframes}
\usage{
page.to.dataframes(page, design)
}
\arguments{
\item{page}{a xml text file the represents a fhir bundle page.}

\item{design}{a structure that specifies which table should contain which entries of the bundle.}
}
\value{
a list of data frames
}
\description{
converts a fhir bundle page to a list of data frames.
design is a named list. Its names are the one of the resulting tables.
The elements of design are lists of 2 elements.
The first one is a XPath expression to locate the entry in a fhir bundle page.
The second one is a named list with XPath expressions of locations to the values of the items in the bundle page.
The names are the column names of the resultung data frames
}
\examples{
page.to.dataframes(
page,
design = list(
MEDICATION = list(
 tag     = ".//Medication",
 items   = list(
	  SYSTEM  = "code/coding/system/@value",
	  CODE    = "code/coding/code/@value",
	  DISPLAY = "code/coding/display/@value"
	) ) ) )
}
