% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{paste_parameters}
\alias{paste_parameters}
\title{Paste Parematers for a FHIR Search Request}
\usage{
paste_parameters(
  parameters = NULL,
  parameters2add = NULL,
  add_question_sign = F
)
}
\arguments{
\item{parameters}{Either a character of length 1 containing an existing parameter string or
a named vector or list of the form
c("_summary" = "count", "gender" = "male") or list("_summary" = "count", "gender" = "female").}

\item{parameters2add}{Either a character of length 1 containing a parameter string to add or
a named vector or list of the form
c("_summary" = "count", "gender" = "male") or list("_summary" = "count", "gender" = "female").}

\item{add_question_sign}{Logical. Should an ?-Sign to be prefixed. Defaults to FALSE.}
}
\value{
A character of length 1 representing the pasted parameter string.
}
\description{
Paste Parematers for a FHIR Search Request
}
\examples{
paste_parameters(
  parameters = "gender=male",
  parameters2add = list("_summary" = "count", "_count" = "1"),
  add_question_sign = TRUE)
}
