% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{paste_paths}
\alias{paste_paths}
\title{Concatenate paths}
\usage{
paste_paths(path1 = "w", path2 = "d", os = "LiNuX")
}
\arguments{
\item{path1}{A string specifying the left hand part of final path.}

\item{path2}{A string specifying the right hand part of final path.}

\item{os}{A string specifying theoperating system you're operating on: windows or linux.}
}
\value{
A string containing the concatenated path.
}
\description{
Concatenates to path string correctly.
}
\examples{
paste_paths("data", "patients")
paste_paths("/data", "patients")
paste_paths("/data/", "patients")
paste_paths("/data", "/patients")
paste_paths("/data/", "/patients/")
paste_paths("data", "patients", "windows")
}
