% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\docType{data}
\name{patient_bundles}
\alias{patient_bundles}
\title{Exemplary FHIR bundles}
\format{
\linkS4class{fhir_bundle_list} containing 2 \linkS4class{fhir_bundle_serialized} objects.
\emph{They have to be unserialized before use, see Examples!}
}
\source{
Generated by:

\code{
bundles <- fhir_search(request="https://server.fire.ly/Patient", max_bundles=2, verbose = 0)

patient_bundles <- fhir_serialize(bundles)
}

\strong{Downloaded 04-26-21}
}
\usage{
patient_bundles
}
\description{
This data example can be used to explore some of the functions from the fhircrackr package
when direct access to a FHIR server is not possible.
}
\details{
patient_bundles is a \linkS4class{fhir_bundle_list} containing \linkS4class{fhir_bundle_serialized} objects
representing FHIR bundles as returned by \code{\link[=fhir_search]{fhir_search()}}.

It contains 2 bundles with Patient resources.
}
\examples{
#unserialize xml objects before doing anything else with them!
patient_bundles <- fhir_unserialize(patient_bundles)


}
\keyword{datasets}
