% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{pluralS}
\alias{pluralS}
\title{Add a 's' to a plural.}
\usage{
pluralS(n)
}
\arguments{
\item{n}{A numeric of length one.}
}
\value{
A character.
}
\description{
Add a 's' to a plural.
}
\examples{
n <- 0
cat(paste0('We have ', n, ' Rabbit', pluralS(n), '.\n'))
n <- n + 1
cat(paste0('Now we have ', n, ' Rabbit', pluralS(n), '.\n'))
n <- n + 1
cat(paste0('Finally we have ', n, ' Rabbit', pluralS(n), '.\n'))
}
