% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rbind_list_of_data_frames}
\alias{rbind_list_of_data_frames}
\title{Bind a list of data frames}
\usage{
rbind_list_of_data_frames(list)
}
\arguments{
\item{list}{A list of data frames to bind}
}
\value{
A single data frame
}
\description{
Rowbinds a list of data frames to one data frame
}
