% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{read.bundle}
\alias{read.bundle}
\title{read.bundle}
\usage{
read.bundle(directory)
}
\arguments{
\item{directory}{the location to store the data.}
}
\value{
the bundle as a list of xml text files.
}
\description{
reads a bundle stored as xml files from a directory
}
\examples{
read.bundle( "result" )
}
