% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{read.bundles}
\alias{read.bundles}
\title{read.bundles}
\usage{
read.bundles(directory)
}
\arguments{
\item{directory}{the location the data are stored.}
}
\value{
the bundles as a list of xml text files.
}
\description{
reads all bundles stored as xml files from a directory
}
\examples{
\dontrun{
bundles.bak <- read.bundles( "result" )
}
}
