% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{result2dfs}
\alias{result2dfs}
\title{result2dfs}
\usage{
result2dfs(result, design, sep = "-+-")
}
\arguments{
\item{result}{a fhir search result as a list of xml text files.}

\item{design}{a structure that specifies which table should contain which entries of the bundle.}

\item{sep}{a string to separate pasted multiple entries}
}
\value{
a list of data frames.
}
\description{
converts all fhir bundles (the result of a get.all.bundles) to a list of data frames
}
\examples{
\dontrun{
result2dfs( bundles, design )
}
}
