% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{sample_identified_resources}
\alias{sample_identified_resources}
\title{Sample Resources of certain IDs}
\usage{
sample_identified_resources(
  endpoint,
  resource,
  ids,
  sample_size = 20,
  seed = as.double(Sys.time()),
  verbose = 1
)
}
\arguments{
\item{endpoint}{A character of length 1 containing the FHIR R4 endpoint.}

\item{resource}{A character of length 1 containing the Resource's name.}

\item{ids}{A character containing the IDs that should be sampled.}

\item{sample_size}{A integer of length 1 containing the size of the sample.}

\item{seed}{A integer of length 1 containing the seed for the random generator.}

\item{verbose}{An integer of length 1 containing the level of verbosity. Defaults to 1.}
}
\value{
A list of bundles containing sampled resources.
}
\description{
Sample Resources of certain IDs
}
\examples{
ids <- get_resources_ids(
  endpoint   = 'https://hapi.fhir.org/baseR4',
  resource   = "Patient",
  parameters = "name=Homer&name=Simpson")
bndls <- sample_identified_resources(
  endpoint    = 'https://hapi.fhir.org/baseR4',
  resource    = "Patient",
  ids         = ids,
  sample_size = 10,
  seed        = 1)
fhir_crack(
  bndls,
  fhir_table_description(
    resource = "Patient",
    cols     = list(
      ID      = "id",
      Homer   = "name/given",
      Simpson = "name/family")))
}
