% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{save_bundles}
\alias{save_bundles}
\title{Save bundles as xml-files}
\usage{
save_bundles(bundles, directory = "result")
}
\arguments{
\item{bundles}{A list of xml objects representing the pages of a fhir bundle.}

\item{directory}{A string containing the path to the folder to store the data in.}
}
\value{
Nothing to return.
}
\description{
Writes all fhir bundles as numbered xml files into a directory.
}
\examples{
\dontrun{
save_bundles(bundles, "result")
}
}
