% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_tree.R
\name{separate_indices}
\alias{separate_indices}
\title{Separate indices from column names}
\usage{
separate_indices(names, bra, ket)
}
\arguments{
\item{names}{A character of length one.}

\item{bra}{A character of length one.}

\item{ket}{A character of length one.}
}
\value{
A character.
}
\description{
Separate indices from column names
}
