% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_tree.R
\name{separate_names}
\alias{separate_names}
\title{Separate names from column names with indices}
\usage{
separate_names(names, bra, ket)
}
\arguments{
\item{names}{A character of length one.}

\item{bra}{A character of length one.}

\item{ket}{A character of length one.}
}
\value{
A character.
}
\description{
Separate names from column names with indices
}
