% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{sort_strings_with_numbers_correctly}
\alias{sort_strings_with_numbers_correctly}
\title{Order Strings with Numbers correctly}
\usage{
sort_strings_with_numbers_correctly(s)
}
\arguments{
\item{s}{A character containing the strings to be sorted.}
}
\value{
A character containing the the given strings sorted.
}
\description{
Order Strings with Numbers correctly
}
\examples{
s <- c("12", "2", "1", "21")
sort(s)
sort_strings_with_numbers_correctly(s = s)
}
