% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{tag.attr}
\alias{tag.attr}
\title{tag.attr}
\usage{
tag.attr(bundle, xpath)
}
\arguments{
\item{bundle}{fhir bundle}

\item{xpath}{the path to the tag's attribute in the xml document.}
}
\value{
a single value or vector
}
\description{
extracts an attribute from tags.
}
\examples{
\dontrun{
tag.attr( bundles[[ 1 ]], xpath = ".//total/@value" )
}
}
