% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{tag_attr}
\alias{tag_attr}
\title{Extract paths}
\usage{
tag_attr(xml, xpath)
}
\arguments{
\item{xml}{A xml document, node, or node set.}

\item{xpath}{A string containing a xpath (1.0) expression}
}
\value{
A character vector containing the attribute.
}
\description{
Extracts an attribute from tags in a xml object.
}
\examples{
\dontrun{
tag_attr(bundles[[1]], xpath = ".//total/@value")
}
}
