% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{write.bundle}
\alias{write.bundle}
\title{write.bundle}
\usage{
write.bundle(bundle, directory)
}
\arguments{
\item{bundle}{a list of xml text files representing the pages of a fhir bundle.}

\item{directory}{the location to store the data.}
}
\value{
nothing to return.
}
\description{
writes a fhir bundle as numbered xml files into a directory.
}
\examples{
write.bundle( my.fhir.xmls, "result" )
}
