% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{write.bundles}
\alias{write.bundles}
\title{write.bundles}
\usage{
write.bundles(bundles, directory = "result")
}
\arguments{
\item{bundles}{a list of xml text files representing the pages of a fhir bundle.}

\item{directory}{the location to store the data.}
}
\value{
nothing to return.
}
\description{
writes all fhir bundle as numbered xml files into a directory.
}
\examples{
\dontrun{
write.bundles( bundles, "result" )
}
}
