% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xtrct_all_columns}
\alias{xtrct_all_columns}
\title{Extract all columns}
\usage{
xtrct_all_columns(child, sep = " -+- ", add_ids = F, xpath = ".//@*")
}
\arguments{
\item{child}{A xml child object, representing one FHIR resource}

\item{sep}{A string to separate pasted multiple entries.}

\item{add_ids}{Logical Scalar. Should indices be added to multiple entries?}

\item{df.columns}{The part of design from \code{\link{fhir_crack}} describing which elements to extract
from the resouce}
}
\description{
Extracts all available values from a single resource
}
