% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{bundle.to.dataframes}
\alias{bundle.to.dataframes}
\title{bundle.to.dataframes}
\usage{
bundle.to.dataframes(bundle, design)
}
\arguments{
\item{bundle}{a list of xml text files representing the pages of a fhir bundle.}

\item{design}{a structure that specifies which table should contain which entries of the bundle.}
}
\value{

}
\description{
converts a fhir bundle to a list of data frames
}
\examples{
bundle.to.dataframes( bundle, design )
}
