% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{conformance}
\alias{conformance}
\title{conformance}
\usage{
conformance(url)
}
\arguments{
\item{url}{the url of the fhir server endpoint.}
}
\value{
a data frame.
}
\description{
get the conformance information about a fhir server.
}
\examples{
\dontrun{
conformance( "https://hapi.fhir.org/baseR4" )
}
}
