% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{download.bundle}
\alias{download.bundle}
\title{download.bundle}
\usage{
download.bundle(fhir.search.request, max.attempts = 5)
}
\arguments{
\item{fhir.search.request}{a fhir search request. it must contain _format=xml.}

\item{max.attempts}{the maximal number of attempts to send a request. Default is 5.}
}
\value{
the downloaded bundle in xml format.
}
\description{
downloads a fhir bundle via fhir search request and return it as xml file.
}
\examples{
\dontrun{
download.bundle( "https://hapi.fhir.org/baseR4/Medication?_format=xml" )
}
}
