% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_body.R
\docType{class}
\name{fhir_body-class}
\alias{fhir_body-class}
\title{an s4 class to represent a body for a post to a fhir server}
\description{
objects of this class should always be created with a call to the function \code{\link[=fhir_body]{fhir_body()}}
}
\section{Slots}{

\describe{
\item{\code{content}}{a length 1 character representing the body for the post}

\item{\code{type}}{a length 1 character defining the type of the body e.g. \code{"application/x-www-form-urlencoded"} or \code{"xml"}}
}}

