% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_body.R
\name{fhir_body}
\alias{fhir_body}
\title{create \linkS4class{fhir_body} object}
\usage{
fhir_body(content, type)
}
\arguments{
\item{content}{a string representing the body for the post in the format specified in \code{type} or
a \linkS4class{fhir_params} object. In the latter case the \code{type} argument should be omitted
and is automatically set to \code{"application/x-www-form-urlencoded"}}

\item{type}{a string defining the type of the body e.g. \code{"application/x-www-form-urlencoded"} or \code{"xml"}}
}
\value{
an object of type \linkS4class{fhir_body}
}
\description{
create \linkS4class{fhir_body} object
}
\examples{
 #body that could be used in a FHIR seach request POSTed to an URL like baseurl/Patient/_search
fhir_body(content = "gender=female&_summary=count", type="application/x-www-form-urlencoded")

#same request providing a fhir_parameters object
fhir_body(content = fhir_parameters(list(c("gender", "female"), c("_summary", "count"))))
}
