% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_bundle.R
\docType{class}
\name{fhir_bundle_serialized-class}
\alias{fhir_bundle_serialized-class}
\title{An S4 class to represent a FHIR bundle in serialized form}
\description{
An S4 class to represent a FHIR bundle in serialized form
}
