% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_bundle.R
\docType{class}
\name{fhir_bundle_xml-class}
\alias{fhir_bundle_xml-class}
\title{An S4 class to represent a FHIR bundle in xml form}
\description{
An S4 class to represent a FHIR bundle in xml form
}
\section{Slots}{

\describe{
\item{\code{next_link}}{A \linkS4class{fhir_url} pointing to the next bundle on the server}

\item{\code{self_link}}{A \linkS4class{fhir_url} pointing to this bundle on the server}
}}

