% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_bundle.R
\name{fhir_bundle_xml}
\alias{fhir_bundle_xml}
\title{Create \linkS4class{fhir_bundle_xml} object}
\usage{
fhir_bundle_xml(bundle)
}
\arguments{
\item{bundle}{A xml-object representing a FHIR bundle}
}
\description{
Create \linkS4class{fhir_bundle_xml} object
}
\examples{
fhir_bundle_xml(xml2::xml_unserialize(patient_bundles[[1]]))

}
