% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{fhir_canonical_design}
\alias{fhir_canonical_design}
\title{Retrieve design of last call to fhir_crack}
\usage{
fhir_canonical_design()
}
\description{
Returns the \linkS4class{fhir_design} of the last call to \code{\link[=fhir_crack]{fhir_crack()}}.
}
\examples{
#load example bundles
bundles <- fhir_unserialize(patient_bundles)

#define design
patients <- fhir_df_description(resource = "Patient")
design <- fhir_design(patients)

result <- fhir_crack(bundles, design)

fhir_canonical_design()

}
