% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_cs}
\alias{fhir_cs}
\title{Get capability statement}
\usage{
fhir_cs(
  url = "https://hapi.fhir.org/baseR4",
  sep = " -+- ",
  remove_empty_columns = T,
  add_indices = F,
  indices_brackets = c("<", ">")
)
}
\arguments{
\item{url}{The url of the FHIR server endpoint.}

\item{sep}{A string to separate pasted multiple entries}

\item{remove_empty_columns}{Logical scalar. Remove empty columns?}

\item{add_indices}{A Logical Scalar.}

\item{indices_brackets}{A Vector of Strings defining the Brackets surrounding the Indices. e.g. c( "<", ">")}
}
\value{
A list of data frames containing the information from the statement
}
\description{
Get the capability statement of a FHIR server.
}
\examples{

cap <- fhir_cs("https://hapi.fhir.org/baseR4")

}
