% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R
\name{fhir_current_request}
\alias{fhir_current_request}
\title{Return FHIR search request used in last call to \code{\link[=fhir_search]{fhir_search()}} or \code{\link[=fhir_url]{fhir_url()}}}
\usage{
fhir_current_request()
}
\value{
An object of class \code{\link[=fhir_url]{fhir_url()}}
}
\description{
Return FHIR search request used in last call to \code{\link[=fhir_search]{fhir_search()}} or \code{\link[=fhir_url]{fhir_url()}}
}
\examples{
\donttest{
request <- fhir_url(url = "https://hapi.fhir.org/baseR4", resource = "Patient")
fhir_current_request()

fhir_search("https://hapi.fhir.org/baseR4/Medication")
fhir_current_request()
}


}
