% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_design.R
\docType{class}
\name{fhir_design-class}
\alias{fhir_design-class}
\title{A S4 class containing a design for \code{\link[=fhir_crack]{fhir_crack()}}}
\description{
A fhir_design is a named list of \linkS4class{fhir_df_desciption} objects. Each df_description
contains information on how to flatten one resource type which will result in one
data.frame. The fhir_design is passed to the function \code{\link[=fhir_crack]{fhir_crack()}} along with a
list of bundles containing fhir resources.
}
\section{Slots}{

\describe{
\item{\code{names}}{The names of the df_descriptions. Those will also be the names of the
resulting data.frames.}
}}

