% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_design.R
\name{fhir_design}
\alias{fhir_design}
\title{Create a \linkS4class{fhir_design} object}
\usage{
fhir_design(..., names)
}
\arguments{
\item{...}{One ore more \linkS4class{fhir_df_description} objects}

\item{names}{The names of the df_descriptions. Those will also be the names of the
resulting data.frames.}
}
\description{
A \linkS4class{fhir_design} is a list of \linkS4class{fhir_df_description} objects and should be created
using the constructor functions for this class. For backwards compatibility
it is for the moment also possible to build it from an old-style design as used in
\verb{fhircrackr (< 1.0.0)}. See examples.
}
\examples{

###Example 1 ####
#create fhir_df_descriptions
df_desc1 <- fhir_df_description(resource = "Patient",
                    cols = c(name = "name/family",
                             gender = "gender",
                             id = "id"),
                    style = fhir_style(sep = "||",
                                       brackets = c("[", "]"),
                                       rm_empty_cols = FALSE
                            )
             )

df_desc2 <- fhir_df_description(resource = "Observation",
                    cols = c("code/coding/system", "code/coding/code")
            )

#create design
#First option
design <- fhir_design(df_desc1, df_desc2, names = c("Patients", "Observations"))

#second option
design <- fhir_design(list(Patients = df_desc1, Observations = df_desc2))

#have a look
design

###Example 2###
#This option will be deprecated at some point

#old style design
old_design <- list(
                 Patients = list(
                    resource = "//Patient",
                    cols = list(
                       name = "name/family",
                       gender = "gender",
                       id = "id"),
                    style = list(
                       sep = "||",
                       brackets = c("[", "]"),
                       rm_empty_cols = FALSE
                    )
                 ),
                 Observations = list(
                    resource = "//Observation",
                    cols = list(
                       code.coding.system = "code/coding/system",
                       code.coding.code = "code/coding/code"
                    )
                 )
              )

new_design <- fhir_design(old_design)
new_design

}
