% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_design.R
\name{fhir_design}
\alias{fhir_design}
\title{Create a \linkS4class{fhir_design} object}
\usage{
fhir_design(..., names)
}
\arguments{
\item{...}{One ore more \linkS4class{fhir_df_description} objects or a named list containing
\linkS4class{fhir_df_description} objects, or an object of class \linkS4class{fhir_df_list}/\linkS4class{fhir_dt_list}.
See examples.}

\item{names}{Optional. The names of the df_descriptions. If no names are provided, the names of the object(s) that
were used in creating the design are taken as the names.}
}
\description{
A \linkS4class{fhir_design} is a named list of \linkS4class{fhir_df_description} objects and should be created
using the function described here. For backwards compatibility
it is for the moment also possible to build it from an old-style design as used in
\verb{fhircrackr (< 1.0.0)}. See examples.
}
\details{
If this function is given an object of class \linkS4class{fhir_df_list} or \linkS4class{fhir_dt_list}, it will
extract the design that was used to create the respective list.
}
\examples{

###Example 1 ####

#create fhir_df_descriptions

pat <- fhir_df_description(resource = "Patient",
                    cols = c(name = "name/family",
                             gender = "gender",
                             id = "id"),
                    style = fhir_style(sep = "||",
                                       brackets = c("[", "]"),
                                       rm_empty_cols = FALSE
                            )
             )

obs <- fhir_df_description(resource = "Observation",
                    cols = c("code/coding/system", "code/coding/code")
            )

#create design

#First option: explicitly define names
fhir_design(pat, obs, names = c("Patients", "Observations"))

#Second option: Names are taken from the df_descriptions
fhir_design(pat, obs))

#Third option: named list
fhir_design(list(Patients = pat, Observations = obs))


###Example 2###
#This option will be deprecated at some point

#old style design
old_design <- list(
                 Patients = list(
                    resource = "//Patient",
                    cols = list(
                       name = "name/family",
                       gender = "gender",
                       id = "id"),
                    style = list(
                       sep = "||",
                       brackets = c("[", "]"),
                       rm_empty_cols = FALSE
                    )
                 ),
                 Observations = list(
                    resource = "//Observation",
                    cols = list(
                       code.coding.system = "code/coding/system",
                       code.coding.code = "code/coding/code"
                    )
                 )
              )

new_design <- fhir_design(old_design)

#have a look at the design
new_design

###Example 3###
#Extract design from fhir_df_list/fhir_dt_list
TODO: Create example when fhir_crack is ready.

}
