% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_df_description.R
\docType{class}
\name{fhir_df_description-class}
\alias{fhir_df_description-class}
\title{A S4 class describing the form of data.frame produced by \code{\link[=fhir_crack]{fhir_crack()}}}
\description{
A fhir_df_description is an element of a \linkS4class{fhir_design} and describes how resources
of one particular resource type should be transformed into a data.frame/data.table.
}
\details{
When the bundle to be cracked contains more than one resource type, each of those
types will be cracked into its own data.frame and needs its own fhir_df_description.
}
\section{Slots}{

\describe{
\item{\code{resource}}{An object of class \linkS4class{fhir_resource_type} defining the resource type that
should be extracted}

\item{\code{cols}}{An object of class \linkS4class{fhir_columns} describing which columns should be created and how.}

\item{\code{style}}{An object of class \linkS4class{fhir_style} describing how to deal with multiple entries and emtpy columns.}
}}

