% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_extract_indices}
\alias{fhir_extract_indices}
\title{Extract indices from molten data.frame with multiple entries}
\usage{
fhir_extract_indices(indexed_data_frame, brackets)
}
\arguments{
\item{indexed_data_frame}{A data frame with indices for multiple entries as produced by \code{\link{fhir_melt_all}}.
Please make sure there is no more than one index present per cell, i.e. all multiple entries have been molten into multiple
rows as achieved by \code{\link{fhir_melt_all}} with the argument \code{rm_indices} set to \code{FALSE}.
All columns have to be of class character.}

\item{brackets}{A string vector of length two defining the brackets that were used in \code{\link{fhir_crack}}.}
}
\value{
A character matrix with same dimensions as \code{indexed_data_frame} containing the indices. For use with
\code{\link{fhir_restore_indices}}.
}
\description{
Extracts a character matrix with indices from a molten indexed data.frame/data.table as created
by \code{\link{fhir_melt_all}} with \code{rm_indices=FALSE}. After extraction, the indices of the data.frame can be removed,
and the data.frame can be manipulated as desired. As long as dimensions as well as row and column
order are preserved during manipulation, the indices can then be restored again with
\code{\link{fhir_restore_indices}}.
}
\examples{
#generate example bundle
bundle <- xml2::read_xml(
"<Bundle>

	<Patient>
		<id value='id1'/>
		<address>
			<use value='home'/>
			<city value='Amsterdam'/>
			<type value='physical'/>
			<country value='Netherlands'/>
		</address>
		<birthDate value='1992-02-06'/>
	</Patient>

	<Patient>
		<id value='id2'/>
		<address>
			<use value='home'/>
			<city value='Rome'/>
			<type value='physical'/>
			<country value='Italy'/>
		</address>
		<address>
			<use value='work'/>
			<city value='Stockholm'/>
			<type value='postal'/>
			<country value='Sweden'/>
		</address>
		<birthDate value='1980-05-23'/>
	</Patient>

	<Patient>
		<id value='id3'/>
		<address>
			<use value='home'/>
			<city value='Berlin'/>
		</address>
		<address>
			<type value='postal'/>
			<country value='France'/>
		</address>
		<address>
			<use value='work'/>
			<city value='London'/>
			<type value='postal'/>
			<country value='England'/>
		</address>
		<birthDate value='1974-12-25'/>
		<birthDate value='1978-11-13'/>
	</Patient>

</Bundle>"
)

#crack fhir resources
dfs <- fhir_crack(bundles = list(bundle),
                 design = list(Patients = list(resource = ".//Patient",
			                                  style= list(brackets = c("[","]"),
			                                   sep="||"))))

#Melt multiple entries
d <- fhir_melt_all(dfs$Patients, brackets = c("[", "]"), sep="||", rm_indices = F)

#Extract indices
fhir_extract_indices(d, brackets = c("[", "]"))
}
