% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_resources.R
\name{fhir_get_resource_ids}
\alias{fhir_get_resource_ids}
\title{Get Resources' IDs}
\usage{
fhir_get_resource_ids(
  base_url,
  resource,
  parameters = NULL,
  username = NULL,
  password = NULL,
  token = NULL,
  cookies = NULL,
  verbose = 0
)
}
\arguments{
\item{base_url}{A character vector of length one specifying the base URL of the FHIR server, e.g. \code{"http://hapi.fhir.org/baseR4"}.}

\item{resource}{A character vector of length one or \linkS4class{fhir_resource_type} object with the resource type to be searched, e.g. \code{"Patient"}.}

\item{parameters}{Optional. Either a length 1 character vector containing properly formatted FHIR search parameters, e.g.
\code{"gender=male&_summary=count"} or a named list or named character vector e.g. \code{list(gender="male", "_summary"="count")}
or \code{c(gender="male", "_summary"="count")}. Note that parameter names beginning with \verb{_} have to be put in quotation marks!}

\item{username}{A character vector of length one containing the username for basic authentication.}

\item{password}{A character vector of length one containing the password for basic authentication.}

\item{token}{A character vector of length one or object of class \link[httr:Token-class]{httr::Token}, for bearer token authentication (e.g. OAuth2). See \code{\link[=fhir_authenticate]{fhir_authenticate()}}
for how to create this.}

\item{cookies}{Optional. A named character vector containing key value pairs for cookies, e.g. \code{c(mycookie = "d385se12394j")}.}

\item{verbose}{An integer of length 1 containing the level of verbosity. Defaults to 0.}
}
\value{
A character vector containing the resource (aka logical) IDs of all requested resources.
}
\description{
Download the resource (aka logical) IDs of all resources matching the FHIR search request build from the resource type and search
parameters specified in \code{resource} and \code{parameters}. This function does not download the entire resources, but only extracts their IDs
using the \verb{_elements} parameter of FHIR Search.
}
\details{
For more information on authentication options, please see the help page of \code{\link[=fhir_search]{fhir_search()}}
}
\examples{
\donttest{
fhir_get_resource_ids(
  base_url   = 'https://vonk.fire.ly/R4',
  resource   = "Patient",
  parameters = "gender=female", verbose=1)
  }
}
\seealso{
\code{\link[=fhir_search]{fhir_search()}}, \code{\link[=fhir_get_resources_by_ids]{fhir_get_resources_by_ids()}}
}
