% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_load}
\alias{fhir_load}
\title{Load bundles from xml-files}
\usage{
fhir_load(directory)
}
\arguments{
\item{directory}{A string containing the path to the folder were the data are stored.}
}
\value{
A list of bundles in xml format.
}
\description{
Reads all bundles stored as xml files from a directory.
}
\examples{
\dontrun{
bundles.bak <- fhir_load("result")
}
}
