% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_melt}
\alias{fhir_melt}
\title{Reconstructring Data Frames}
\usage{
fhir_melt(indexed_data_frame, brackets = c("<", ">"), sep = " -+- ")
}
\arguments{
\item{indexed_data_frame}{A Data Frame with indexed multiple Entries in its Columns.}

\item{brackets}{A Cahracter Vector of Length 2, holding the Brackets.}

\item{sep}{A Character Scalar, the Separator.}
}
\value{
A Data Frame.
}
\description{
Reconstructring Data Frames
}
\examples{
\dontrun {
fhir_melt( df )
}

}
