% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_melt}
\alias{fhir_melt}
\title{Reconstructring Data Frames}
\usage{
fhir_melt(
  indexed_data_frame,
  column.prefix = "id",
  brackets = c("<", ">"),
  sep = " -+- ",
  id.name = "#-#+#ID#+#-#"
)
}
\arguments{
\item{indexed_data_frame}{A Data Frame with indexed multiple entries in its columns.}

\item{column.prefix}{A String specifiying the column(s) to be melted.}

\item{brackets}{A character vector of Length 2, holding the Brackets.}

\item{sep}{A string, the separator.}
}
\value{
A data frame without indices
}
\description{
Reconstructring Data Frames
}
\examples{
\dontrun{
fhir_melt( df )
}

}
