% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrack.R
\name{fhir_next_bundle_url}
\alias{fhir_next_bundle_url}
\title{Last Next Link}
\usage{
fhir_next_bundle_url()
}
\value{
The next link.
}
\description{
Returns the next link of the recently downloaded bundle.
}
