% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_parameters.R
\docType{class}
\name{fhir_parameters-class}
\alias{fhir_parameters-class}
\title{An S4 class to represent parameters for FHIR search}
\description{
Objects of this class should always be created with a call to \code{\link[=fhir_parameters]{fhir_parameters()}}
}
\section{Slots}{

\describe{
\item{\code{names}}{The names (keys) of the search parameters.}
}}

