% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_parameters.R
\name{fhir_parameters}
\alias{fhir_parameters}
\title{Create \linkS4class{fhir_parameters} object}
\usage{
fhir_parameters(parameters)
}
\arguments{
\item{parameters}{Either a length 1 character containing properly formatted FHIR search parameters, e.g.
\code{"gender=male&_summary=count"} or a named list or named character vector e.g. \code{list(gender="male", "_summary"="count")}
or \code{c(gender="male", "_summary"="count")}. Note that parameters beginning with \verb{_} have to be put in quotation marks!}
}
\description{
A \linkS4class{fhir_parameters} object can be created in three ways:
\itemize{
\item You provide a length 1 character with all the search parameters pasted together properly
\item You provide a named character vector with names representing the keys and values
representing the values of the search parameters
\item You provide a named list of strings with names representing the keys and values
representing the values of the search parameters
}
}
\examples{
#two ways to create the same fhir_parameters object

#using a string
fhir_parameters("gender=male&birthdate=le2000-01-01&_summary=count")

#using a character vector
fhir_parameters(c(gender="male", birthdate="le2000-01-01", "_summary"="count"))

#using a list
fhir_parameters(list(gender="male", birthdate="le2000-01-01", "_summary"="count"))
}
