% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_resource_type.R
\name{fhir_resource_type}
\alias{fhir_resource_type}
\title{Create \linkS4class{fhir_resource_type} object}
\usage{
fhir_resource_type(string)
}
\arguments{
\item{string}{A string containing the resource type. Will mostly be one of the official FHIR resource
types listed at https://hl7.org/FHIR/resourcelist.html}
}
\value{
An fhir_resource object
}
\description{
This function creates an object of class \link{fhir_resource_type_class} by taking a string defining a FHIR resource type
and formating it correctly,i.e. removing white space and slashes. It also checks the resource against the list
of resources provided at https://hl7.org/FHIR/resourcelist.html and converts cases appropriately
if the resource is found in that list.
}
\examples{
fhir_resource("MedicationAdministration")
}
