% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_resource.R
\docType{methods}
\name{fhir_resource_xml}
\alias{fhir_resource_xml}
\alias{fhir_resource_xml,xml_document-method}
\alias{fhir_resource_xml,xml_node-method}
\alias{fhir_resource_xml,character-method}
\title{Create \linkS4class{fhir_resource_xml} object}
\usage{
fhir_resource_xml(resource)

\S4method{fhir_resource_xml}{xml_document}(resource)

\S4method{fhir_resource_xml}{xml_node}(resource)

\S4method{fhir_resource_xml}{character}(resource)
}
\arguments{
\item{resource}{A xml-object representing a FHIR resource}
}
\description{
Create \linkS4class{fhir_resource_xml} object
}
\examples{
fhir_resource_xml(resource = xml2::read_xml("<Patient><id value = '1'/></Patient>"))

}
