% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_entries.R
\name{fhir_restore_indices}
\alias{fhir_restore_indices}
\title{Restore indices to molten data.frame/data.table with multiple entries}
\usage{
fhir_restore_indices(d, index_matrix)
}
\arguments{
\item{d}{A data.frame/data.table as produced by \code{\link[=fhir_melt_all]{fhir_melt_all()}} without indices.}

\item{index_matrix}{A character matrix with the same dimensions as \code{d}, as produced by \code{\link[=fhir_extract_indices]{fhir_extract_indices()}}}
}
\value{
\code{d} but with the indices from \code{index_matrix}.
}
\description{
Takes a molten data.frame/data.table containing no indices as produced by \code{\link[=fhir_melt_all]{fhir_melt_all()}} and
restores its indices from a matrix of indices as produces by \code{\link[=fhir_extract_indices]{fhir_extract_indices()}}. Dimensions as
well as row and column ordering of index matrix and data frame have to be identical!
}
\examples{
#unserialize example bundles
bundles <- fhir_unserialize(example_bundles2)

#crack fhir resources
patients <- fhir_table_description(resource = "Patient",
                               style = fhir_style(brackets = c("[","]"),
			                                     sep="||")
			                  )
df <- fhir_crack(bundles = bundles,
                 design = patients)

#Melt multiple entries
d <- fhir_melt_all(indexed_data_frame = df, brackets = c("[", "]"), sep="||", rm_indices = FALSE)

#Extract indices
indices <- fhir_extract_indices(indexed_data_frame = d, brackets = c("[", "]"))

#remove indices
d_removed <- fhir_rm_indices(indexed_data_frame = d, brackets = c("[", "]"))

#restore indices
d_restored <- fhir_restore_indices(d = d_removed, indices)

#compare
identical(d, d_restored)
}
\seealso{
\code{\link[=fhir_melt]{fhir_melt()}}, \code{\link[=fhir_melt_all]{fhir_melt_all()}}, \code{\link[=fhir_rm_indices]{fhir_rm_indices()}}
}
