% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R
\name{fhir_save}
\alias{fhir_save}
\title{Save FHIR bundles as xml-files}
\usage{
fhir_save(bundles, directory)
}
\arguments{
\item{bundles}{A list of xml objects representing the FHIR bundles.}

\item{directory}{A character vector of length one containing the path to the folder to store the data in.}
}
\description{
Writes a list of FHIR bundles as numbered xml files into a directory.
}
\examples{
#unserialize example bundle
bundles <- fhir_unserialize(medication_bundles)

#save all bundles to temporary directory
fhir_save(bundles, directory = tempdir())

#save only two bundles (the second and the third) to temporary directory
fhir_save(bundles[c(2,3)], directory = tempdir())
}
