% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_save}
\alias{fhir_save}
\title{Save bundles as xml-files}
\usage{
fhir_save(bundles, directory = "result")
}
\arguments{
\item{bundles}{A list of xml objects representing the pages of a fhir bundle.}

\item{directory}{A string containing the path to the folder to store the data in.}
}
\value{
Nothing to return.
}
\description{
Writes all fhir bundles as numbered xml files into a directory.
}
\examples{
\dontrun{
fhir_save(bundles, "result")
}
}
