% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{fhir_search}
\alias{fhir_search}
\title{Download fhir search result}
\usage{
fhir_search(
  request,
  username = NULL,
  password = NULL,
  max.bundles = Inf,
  verbose = T,
  max.attempts = 5,
  delay.between.attempts = 10
)
}
\arguments{
\item{max.bundles}{Maximal number of bundles to get. Defaults to Inf meaning all available bundles are downloaded.}
}
\value{
A list of bundles in xml format.
}
\description{
Downloads all fhir bundles of a fhir search request from a fhir server.
}
\examples{
\dontrun{
bundles <- fhir_search("https://hapi.fhir.org/baseR4/Medication?", max.bundles=10)
}
}
