% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\name{fhir_search}
\alias{fhir_search}
\title{Download fhir search result}
\usage{
fhir_search(
  request,
  username = NULL,
  password = NULL,
  max_bundles = Inf,
  verbose = 1,
  max_attempts = 5,
  delay_between_attempts = 10
)
}
\arguments{
\item{request}{A string containing the full FHIR search request.}

\item{username}{A string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{A string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}

\item{max_bundles}{Maximal number of bundles to get. Defaults to Inf meaning all available bundles are downloaded.}

\item{verbose}{A integer scalar. Level of downloading information to be printed to the console? Defaults to 1.}

\item{max_attempts}{A numeric scalar. The maximal number of attempts to send a request, defaults to 5.}

\item{delay_between_attempts}{A numeric scalar specifying the delay in seconds between two attempts. Defaults to 10.}
}
\value{
A list of bundles in xml format.
}
\description{
Downloads all fhir bundles of a fhir search request from a fhir server.
}
\examples{
bundles <- fhir_search("https://hapi.fhir.org/baseR4/Medication?", max_bundles=3)
}
