% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R
\name{fhir_search}
\alias{fhir_search}
\title{Download FHIR search result}
\usage{
fhir_search(
  request = fhir_current_request(),
  username = NULL,
  password = NULL,
  max_bundles = Inf,
  verbose = 1,
  max_attempts = 5,
  delay_between_attempts = 10,
  log_errors = NULL,
  save_to_disc = NULL,
  delay_between_pages = 0,
  directory = paste0("FHIR_bundles_", gsub("-| |:", "", Sys.time()))
)
}
\arguments{
\item{request}{An object of class \linkS4class{fhir_url} or a string containing the full FHIR search request. It is
recommended to explicitly create the request via \code{\link[=fhir_url]{fhir_url()}} as this will do some validity checks and format the url properly.
Defaults to \code{\link[=fhir_current_request]{fhir_current_request()}}}

\item{username}{A string containing the username for basic authentication. Defaults to NULL, meaning no authentication.}

\item{password}{A string containing the password for basic authentication. Defaults to NULL, meaning no authentication.}

\item{max_bundles}{Maximal number of bundles to get. Defaults to Inf meaning all available bundles are downloaded.}

\item{verbose}{An Integer Scalar.  If 0, nothings is printed, if 1, only finishing message is printed, if > 1,
downloading progress will be printed. Defaults to 2.}

\item{max_attempts}{A numeric scalar. The maximal number of attempts to send a request, defaults to 5.}

\item{delay_between_attempts}{A numeric scalar specifying the delay in seconds between two attempts. Defaults to 10.}

\item{log_errors}{Either \code{NULL} or a string indicating the name of a file in which to save the http errors.
\code{NULL} means no error logging. When a file name is provided, the errors are saved in the specified file. Defaults to \code{NULL}}

\item{save_to_disc}{Either \code{NULL} or a string indicating the name of a directory in which to save the bundles.
If a directory name is provided, the bundles are saved as numerated xml-files into the directory specified
and not returned as a bundle list in the R session. This is useful when a lot of bundles are to be downloaded and keeping them all
in one R session might overburden working memory. When the download is complete, the bundles can be loaded into R using \code{\link[=fhir_load]{fhir_load()}}.
Defaults to \code{NULL}, i.e. bundles are returned as a list within the R session.}

\item{delay_between_pages}{A numeric scalar specifying a time in seconds to wait between pages of the search result,
i.e. between downloading the current bundle and the next bundle. This can be used to avoid choking a weak server with
too many requests to quickly. Defaults to zero.}

\item{directory}{Deprecated. Please specify the directory directly in the \code{save_to_disc} argument.}
}
\value{
A \linkS4class{fhir_bundle_list} when\code{save_to_disc = NULL} (the default),  else \code{NULL}.
}
\description{
Downloads all FHIR bundles of a FHIR search request from a FHIR server.
}
\examples{
\donttest{

#create fhir search url
request <- fhir_url(url = "https://server.fire.ly",
                    resource = "Patient",
                    parameters = c(gender="female"))
#download bundles
bundles <- fhir_search(request, max_bundles = 5)
}
}
