% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_url.R
\name{fhir_search_url}
\alias{fhir_search_url}
\title{Create URL for FHIR search}
\usage{
fhir_search_url(base, resource, parameters)
}
\arguments{
\item{base}{A character of length 1 with the base URL to the FHIR server, e.g. \code{"http://hapi.fhir.org/baseR4"}}

\item{resource}{A character of length 1 with the resource type to be searched, e.g. \code{"Patient"}}

\item{parameters}{Optional. An object of type \linkS4class{fhir_parameters} or a string or list that can be
converted into an \linkS4class{fhir_parameters} object. See the help for \code{\link[=fhir_parameters]{fhir_parameters()}} for more info.}
}
\description{
This function creates an object of class \linkS4class{fhir_search_url} which represents a URL for
a FHIR search request. A valid URL must contain a base URL and a resource type and can contain additional
search parameters. For more info on FHIR search see https://www.hl7.org/fhir/search.html
}
\examples{

fhir_search_url(
   base = "http://hapi.fhir.org/baseR4",
   resource = "Patient"
 )

fhir_search_url(
   base = "http://hapi.fhir.org/baseR4",
   resource = "Patient",
   parameters = "gender=male&_summary=count"
 )

fhir_search_url(
   base = "http://hapi.fhir.org/baseR4",
   resource = "Patient",
   parameters = list(c("gender", "male"),
                     c("_summary", "count")
                     )
 )

fhir_search_url(
   base = "http://hapi.fhir.org/baseR4",
   resource = "Patient",
   parameters = fhir_parameters(
                  fhir_key_value_pair("gender", "male"),
                  fhir_key_value_pair("_summary", "count")
                )
 )

fhir_search_url(
   base = "http://hapi.fhir.org/baseR4",
   resource = "Patient",
   parameters = list(fhir_key_value_pair("gender", "male"),
                     fhir_key_value_pair("summary", "count")
                     )
 )
}
