% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_style.R
\docType{class}
\name{fhir_style-class}
\alias{fhir_style-class}
\title{An S4 class to represent a design for cracking FHIR resources}
\description{
An S4 class to represent a design for cracking FHIR resources
}
\section{Slots}{

\describe{
\item{\code{sep}}{A string to separate pasted multiple entries. Defaults to \code{":::"}.}

\item{\code{brackets}}{A character vector of length two defining the brackets surrounding indices for multiple entries,
e.g. \code{c("<|", "|>")}. If this is empty (i.e. character of length 0, the default), no indices will be added to multiple entries.
Empty strings (\code{""}) are not allowed.}

\item{\code{rm_empty_cols}}{Logical scalar. Remove empty columns? Defaults to FALSE.}
}}

