% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_style.R
\name{fhir_style}
\alias{fhir_style}
\title{Create fhir_style object}
\usage{
fhir_style(sep = " ", brackets = character(), rm_empty_cols = TRUE)
}
\arguments{
\item{sep}{A character vector of length 1 to separate pasted multiple entries. Defaults to \code{" "}}

\item{brackets}{A character vector of length two defining the brackets surrounding indices for multiple entries, e.g. \code{c( "<", ">")}.
If this is empty (i.e. character of length 0, the default) or 'NULL', no indices will be added to multiple entries. If it is a character
vector of length 1, it will be recycled to length two, i.e. \code{"|"} will become \code{c("|", "|")}.}

\item{rm_empty_cols}{Logical scalar. Remove empty columns? Defaults to \code{TRUE}.}
}
\value{
A fhir_style object
}
\description{
This function creates an object of class \code{fhir_style}.
It contains the three elements \code{sep}, \code{brackets} and \code{rm_empty_cols}. See Details.
}
\details{
A \code{fhir_style} object is part of a \code{fhir_df_description} which in turn is part of a \code{fhir_design} and
ultimately used in \code{\link[=fhir_crack]{fhir_crack()}}. A \code{fhir_style} object contains three elements:
\itemize{
\item \code{sep}: A string defining the separator used to separate multiple entries for the same element in a FHIR resource,
e.g. multiple \code{address/city} elements in a Patient resource.
\item \code{brackets}: A character vector of length two defining the brackets surrounding indices for multiple entries, e.g. \code{c( "<", ">")}.
If this is empty (i.e. character of length 0, the default), no indices will be added to multiple entries.
\item \code{rm_empty_cols}: A logical scalar defining whether or not to remove empty columns after cracking. Empty columns arise when you
try to extract an element that doesn't appear in any of the resources.
A \code{fhir_style} object looks for example like this:
}\preformatted{sep: ' '
brackets: '[' ']'
rm_empty_cols: FALSE
}
}
\examples{
fhir_style(sep = " ",
           brackets = c("[", "]"),
           rm_empty_cols = FALSE)

}
