% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_style.R
\name{fhir_style}
\alias{fhir_style}
\title{Create fhir_style object}
\usage{
fhir_style(sep = " ", brackets = character(), rm_empty_cols = TRUE)
}
\arguments{
\item{sep}{A character vector of length 1 to separate pasted multiple entries. Defaults to \code{" "}}

\item{brackets}{A character vector of length two defining the brackets surrounding indices for multiple entries, e.g. \code{c( "<", ">")}.
If this is empty (i.e. character of length 0, the default) or 'NULL', no indices will be added to multiple entries.}

\item{rm_empty_cols}{Logical scalar. Remove empty columns? Defaults to \code{TRUE}.}
}
\value{
A fhir_style object
}
\description{
This function creates an object of class \code{fhir_style}. It contains the three elements \code{sep}, \code{brackets} and \code{rm_empty_cols}
}
\examples{
fhir_style(sep = " ", brackets = c("[", "]"), rm_empty_cols = FALSE)
}
