% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R
\name{fhir_unserialize}
\alias{fhir_unserialize}
\title{Unserialize a \linkS4class{fhir_bundle} or \linkS4class{fhir_bundle_list}}
\usage{
fhir_unserialize(bundles)
}
\arguments{
\item{bundles}{A \linkS4class{fhir_bundle} or \linkS4class{fhir_bundle_list} object.}
}
\value{
A  \linkS4class{fhir_bundle_serialized} or \linkS4class{fhir_bundle_list} object.
}
\description{
Unserializes a list of FHIR bundles that have been serialized to allow for saving in .rda or .RData format.
}
\examples{
bundles <- fhir_unserialize(medication_bundles)

}
