% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_resources.R
\name{fhir_update_request}
\alias{fhir_update_request}
\title{Update the current FHIR search request}
\usage{
fhir_update_request(..., append = FALSE, return_request = TRUE)
}
\arguments{
\item{...}{calls to \code{\link{fhir_key_value}}}

\item{append}{Logical. Keep key value pairs from current search request?
Defaults to \code{FALSE}, meaning only base url and resource type from current request are kept.
If \code{TRUE}, the new key value pairs will be added to the existing ones.}

\item{return_request}{Logical. Return string with updated request? Defaults to \code{TRUE}.}
}
\value{
A string with the updated FHIR search request or \code{NULL}.
}
\description{
Takes the current request (the search request URL from either the last call to
\code{\link{fhir_search}} or \code{\link{fhir_build_request}}) an updates the search
parameters with new calls to \code{\link{fhir_key_value}}. The updated request can be
accessed with \code{\link{fhir_current_request}}.
}
\examples{
#build request
fhir_build_request(fhir_base("http://hapi.fhir.org/baseR4"),
               fhir_resource("Patient"),
               fhir_key_value(key = "gender", value = "female"))

#access current request
fhir_current_request()

#update and keep former key value pairs
fhir_update_request(fhir_key_value(key = "_count", value = "10"), append=TRUE)
fhir_current_request()

#update and replace former key value pairs
fhir_update_request(fhir_key_value(key = "gender", value = "male"),
append = FALSE, return_request = TRUE)

}
