% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_xpath_expression.R
\name{fhir_xpath_expression}
\alias{fhir_xpath_expression}
\title{Create fhir_xpath_expression}
\usage{
fhir_xpath_expression(expression)
}
\arguments{
\item{expression}{A string with the XPath expression}
}
\value{
A XPath expression object
}
\description{
This function takes a string, checks whether it is a valid XPath expression
and returns it as an fhir_xpath_expression object
}
\examples{
fhir_xpath_expression("//Patient")

fhir_xpath_expression("name/given")

}
