% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhir_xpath_expression.R
\name{fhir_xpath_expression}
\alias{fhir_xpath_expression}
\title{Create fhir_xpath_expression}
\usage{
fhir_xpath_expression(expression)
}
\arguments{
\item{expression}{A character vector of the XPath expressions}
}
\value{
A XPath expression object
}
\description{
This function takes a character vector, checks whether it contains valid XPath (1.0) expressions
and returns it as an fhir_xpath_expression object. These objects are used in \code{fhir_parameters} objects.
}
\examples{
fhir_xpath_expression(c("//Patient", "name/given"))

}
