% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{get.bundle}
\alias{get.bundle}
\title{get.bundle}
\usage{
get.bundle(request, username = NULL, password = NULL, max.attempts = 5)
}
\arguments{
\item{request}{an request for a fhir bundle. it must contain _format=xml.}

\item{username}{a string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{a string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}

\item{max.attempts}{the maximal number of attempts to send a request. Default is 5.}
}
\value{
the downloaded bundle in xml format.
}
\description{
downloads a single fhir bundle via fhir search request and return it as xml file.
}
\examples{
\dontrun{
get.bundle( request = "https://hapi.fhir.org/baseR4/Medication?_count=500&_format=xml" )
}
}
