% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{get.bundles}
\alias{get.bundles}
\title{get.bundles}
\usage{
get.bundles(
  request,
  username = NULL,
  password = NULL,
  max.bundles = Inf,
  verbose = T,
  max.attempts = 5,
  delay.between.attempts = 10
)
}
\arguments{
\item{request}{A string containing the full fhir search request.}

\item{username}{A string containing the username for basic authentification. Defaults to NULL, meaning no authentification.}

\item{password}{A string containing the passwort for basic authentification. Defaults to NULL, meaning no authentification.}

\item{max.bundles}{Maximal number of bundles to get. Defaults to Inf meaning all available bundles are downloaded.}

\item{verbose}{A logical scalar. Should downloading information be printed to the console? Defaults to TRUE.}

\item{max.attempts}{A numeric scalar. The maximal number of attempts to send a request, defaults to 5.}

\item{delay.between.attempts}{A scalar numeric specifying the delay in seconds between two attempts. Defaults to 10.}
}
\value{
A list of bundles in xml format.
}
\description{
Downloads all fhir bundles of a fhir search request from a fhir server.
}
\examples{
\dontrun{
bundles <- get.bundles( "https://vonk.fire.ly/R4/Medication?_format=xml" )
}
}
