% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{load.bundles}
\alias{load.bundles}
\title{load.bundles}
\usage{
load.bundles(directory)
}
\arguments{
\item{directory}{the location the data are stored.}
}
\value{
the bundles as a list of xml text files.
}
\description{
reads all bundles stored as xml files from a directory
}
\examples{
\dontrun{
bundles.bak <- load.bundles( "result" )
}
}
