% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{lst}
\alias{lst}
\title{lst}
\usage{
lst(..., prefix = NULL, suffix = NULL)
}
\arguments{
\item{...}{a scalar or a vector of scalars.}

\item{prefix}{a prefix in front of the result text.}

\item{suffix}{a suffix after the result text.}
}
\value{
a named list, where the names are the content surrounded bei prefix and suffix
}
\description{
transforms a vector of items to a named list. The names are created the a prefix and a suffix surrounding the items.
}
\examples{
lst( LETTERS[ 1 : 5 ], pre = "id[", post = "]" )
}
