% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhircrackr.R
\docType{data}
\name{medication_bundles}
\alias{medication_bundles}
\title{Exemplary FHIR bundles}
\format{
List of length 3 containing \emph{serialized} "xml_document" objects, each representing one bundle from a
FHIR search request. \emph{They have to be unserialized before use, see Usage!}
}
\source{
\url{https://hapi.fhir.org/baseR4/MedicationStatement?code=http://snomed.info/ct|429374003&_include=MedicationStatement:subject}
[Downloaded 06-22-20]
}
\usage{
#unserialize xml objects before doing anything else with them!
medication_bundles <- lapply(medication_bundles, xml2::xml_unserialize)
}
\description{
This data example can be used to explore some of the functions from the fhircrackr package
when direct access to a FHIR server is not possible.
}
\details{
medication_bundles is a list of \emph{serialized} xml objects representing FHIR bundles as returned by fhir_search().
It contains 3 bundles with MedicationStatement resources representing Medications with Snomed CT code
429374003 and the respective Patient resources that are linked to these MedicationStatements.

It corresponds to Example 2 of downloading and flattening FHIR resources from the vignette of the package.
}
\keyword{datasets}
