% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{order_strings_with_numbers_correctly}
\alias{order_strings_with_numbers_correctly}
\title{Order Strings with Numbers correctly}
\usage{
order_strings_with_numbers_correctly(s)
}
\arguments{
\item{s}{A character vector containing the strings to be ordered.}
}
\value{
A character vector containing the order of the given strings.
}
\description{
Order Strings with Numbers correctly
}
\examples{
s <- c('[1.0]', '[10.1]', '[2.0]')
order(s)
order_strings_with_numbers_correctly(s = s)
}
