% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{paste.paths}
\alias{paste.paths}
\title{paste.paths}
\usage{
paste.paths(path1 = "w", path2 = "d", os = "LiNuX")
}
\arguments{
\item{path1}{the left hand part of final path.}

\item{path2}{the right hand part of final path.}

\item{os}{the operating system you'r operating on: windows or linux.}
}
\value{
a string
}
\description{
concatenates to paths strings correctly
}
\examples{
\dontrun{
paste.paths( "data", "patients" )
paste.paths( "/data", "patients" )
paste.paths( "/data/", "patients" )
paste.paths( "/data", "/patients" )
paste.paths( "/data/", "/patients/" )
paste.paths( "data", "patients", "windows" )
}
}
