% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{save.bundles}
\alias{save.bundles}
\title{save.bundles}
\usage{
save.bundles(bundles, directory = "result")
}
\arguments{
\item{bundles}{A list of xml objects representing the pages of a fhir bundle.}

\item{directory}{A string containing the path to the folder to store the data in.}
}
\value{
Nothing to return.
}
\description{
Writes all fhir bundles as numbered xml files into a directory.
}
\examples{
\dontrun{
save.bundles( bundles, "result" )
}
}
