% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{sort_strings_with_numbers_correctly}
\alias{sort_strings_with_numbers_correctly}
\title{Sort Strings with Numbers correctly}
\usage{
sort_strings_with_numbers_correctly(s)
}
\arguments{
\item{s}{A character vector containing the strings to be sorted.}
}
\value{
A character vector containing the the given strings sorted.
}
\description{
Sort Strings with Numbers correctly
}
\examples{
s <- c('[1.0]', '[10.1]', '[2.0]')
sort(s)
sort_strings_with_numbers_correctly(s = s)
}
