% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxml4.R
\name{tag.attr}
\alias{tag.attr}
\title{tag.attr}
\usage{
tag.attr(xml, xpath)
}
\arguments{
\item{xml}{a document, node, or node set}

\item{xpath}{a string containing a xpath (1.0) expression}
}
\value{
a single value or vector
}
\description{
extracts an attribute from tags.
}
\examples{
\dontrun{
tag.attr( bundles[[ 1 ]], xpath = ".//total/@value" )
}
}
