% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{th}
\alias{th}
\title{th}
\usage{
th(n)
}
\arguments{
\item{n}{a number or a vector of numbers.}
}
\value{
a vector of strings
}
\description{
add the right suffix to a number or a vector of numbers. e.g. 1st 2nd 3rd ...
}
\examples{
\dontrun{
th( 0 : 4 )
}
}
