% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{xml2df}
\alias{xml2df}
\title{xml2df}
\usage{
xml2df(xml, dsgn.df, sep = " -+- ")
}
\arguments{
\item{xml}{xml doc or xml node}

\item{dsgn.df}{a design for a single data frame}

\item{sep}{a string to separate pasted multiple entries}
}
\value{
a data frame
}
\description{
converts a xml doc or xml node to one data frame
}
\examples{
\dontrun{
xml2df( xml, design$Patient )
}
}
