% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fhi.R
\name{xml2df}
\alias{xml2df}
\title{xml2df}
\usage{
xml2df(xml, dsgn.df, sep = " -+- ")
}
\arguments{
\item{xml}{An xml doc or xml node object.}

\item{dsgn.df}{A design for a single data frame.}

\item{sep}{A string used to separate pasted multiple entries}
}
\value{
A data frame containing the data specified in \code{dsgn.df}.
}
\description{
Converts an xml doc or xml node object to one data frame.
}
\examples{
\dontrun{
xml2df( xml, design$Patient )
}
}
