% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xtrct_all_columns}
\alias{xtrct_all_columns}
\title{Extract all columns}
\usage{
xtrct_all_columns(child, sep = " -+- ", add_ids = F, xpath = ".//@*")
}
\arguments{
\item{child}{A xml child object, representing one FHIR resource
from the resouce}

\item{sep}{A String to separate pasted multiple entries.}

\item{add_ids}{Logical Scalar. Should indices be added to multiple entries?}

\item{xpath}{A String to locate data in tree via xpath.}
}
\description{
Extracts all available values from a single resource
}
