% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xtrct_all_columns}
\alias{xtrct_all_columns}
\title{Extract all columns}
\usage{
xtrct_all_columns(
  child,
  sep = " -+- ",
  xpath = ".//@*",
  add_indices = F,
  brackets = c("<", ">")
)
}
\arguments{
\item{child}{A xml child object, representing one FHIR resource
from the resouce}

\item{sep}{A String to separate pasted multiple entries.}

\item{xpath}{A String to locate data in tree via xpath.}

\item{add_indices}{A Logical Scalar.}

\item{brackets}{A Vector of Strings defining the Brackets surrounding the Indices. e.g. c( "<", ">")}
}
\description{
Extracts all available values from a single resource
}
\examples{
#unserialize example bundle
bundles <- fhir_unserialize(medication_bundles)

#extract child
child <- xml2::xml_find_first(bundles[[1]], ".//MedicationStatement")

#Extract all columns
result <- fhircrackr:::xtrct_all_columns(child)

}
