% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xtrct_columns}
\alias{xtrct_columns}
\title{Extract columns}
\usage{
xtrct_columns(child, df.columns, sep = " -+- ", add_ids = F)
}
\arguments{
\item{child}{A xml child object, representing one FHIR resource}

\item{df.columns}{The part of design from \code{\link{fhir_crack}} describing which elements to extract
from the resouce}

\item{sep}{A string to separate pasted multiple entries.}

\item{add_ids}{Logical Scalar. Should indices be added to multiple entries?}
}
\description{
Extracts defined values from a single resource
}
\examples{
#unserialize example bundle
bundles <- fhir_unserialize(medication_bundles)

#extract child
child <- xml2::xml_find_first(bundles[[1]], ".//MedicationStatement")

#define columns
cols <-list(
	SYSTEM  = "medicationCodeableConcept/coding/system/@value",
	CODE    = "medicationCodeableConcept/coding/code/@value",
	DISPLAY = "medicationCodeableConcept/coding/display/@value"
)

#Extract columns
result <- fhircrackr:::xtrct_columns(child, cols)
}
